config: the rest of the resolution pipeline moves out of the record
Seventeen handlers, the last of the bodies: `model_hook.py` (the per-model and
per-capability adjustments, the mamba radix cache, the language-model-only
rule), `memory_hook.py` (the GPU memory budget), `dllm_hook.py` (diffusion-LM
inference), plus the deprecated-argument mapping, the environment variables,
the remaining validations, the missing-value defaults, the hidden-states mode,
the prefill delayer compatibility, tokenizer batching and the multimodal
feature transport joining `serving_hook.py`; the load format joining
`model_path_hook.py`; the encoder disaggregation joining
`pd_disaggregation_hook.py`; and deterministic inference joining
`attention_hook.py`.

Four bodies stay, and none is a pipeline stage: `_resolved` and
`_resolved_attention_backends` are the record's own read wrappers that moved
handlers call back into, `_parsed_modelexpress_config` memoizes on the record,
and `_apply_cuda_graph_disaggregation_roles` reads the record's lock set.

`ServerArgs` is 6835 lines, from 11331: 483 field declarations, the CLI, the
dispatcher, and a slot per handler.

A slot keeps the method's own signature.
`_handle_mamba_radix_cache(self, model_arch)` is the only handler that takes an
argument, and a slot written as `(self)` would drop it.

Two module-level names stay on the record and are read where they are used, so
no hook module imports the record at module scope:
`RADIX_SUPPORTED_DETERMINISTIC_ATTENTION_BACKEND` and
`resolve_encoder_transfer_backend`.

62 launch shapes report the same outcome as before, and the 24-shape resolution
dump is field-for-field identical.
