+++ python/sglang/srt/arg_groups/memory_hook.py
@@ -1,0 +1,268 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Server-argument resolution for the GPU memory budget."""
+
+from __future__ import annotations
+
+import copy
+import logging
+from typing import Any
+
+from sglang.srt.arg_groups.overrides import (
+    declare_resolution,
+    resolving_view,
+)
+from sglang.srt.environ import envs
+
+logger = logging.getLogger(__name__)
+
+
+def handle_gpu_memory_settings(server_args: Any, gpu_mem):
+    """
+    Configure GPU memory-dependent settings including
+    chunked_prefill_size, cuda_graph_config[decode].max_bs, and mem_fraction_static.
+
+    Here are our heuristics:
+    - Set chunked_prefill_size and cuda_graph_config[decode].max_bs based on the GPU memory capacity.
+      This is because GPUs with more memory are generally more powerful, we need to use a larger
+      chunked_prefill_size and a larger decode max_bs to fully utilize the GPU.
+    - Then set mem_fraction_static based on chunked_prefill_size and decode max_bs.
+
+      GPU memory capacity = model weights + KV cache pool + activations + cuda graph buffers
+
+      The argument mem_fraction_static is defined as (model weights + KV cache pool) / GPU memory capacity,
+      or equivalently, mem_fraction_static = (GPU memory capacity - activations - cuda graph buffers) / GPU memory capacity.
+
+      In order to compute mem_fraction_static, we need to estimate the size of activations and cuda graph buffers.
+      The activation memory is proportional to the chunked_prefill_size.
+      The cuda graph memory is proportional to the decode max_bs.
+      We use reserved_mem = chunked_prefill_size * 1.5 + max_bs * 2 to estimate the size of activations and cuda graph buffers in GB,
+      and set mem_fraction_static = (GPU memory capacity - reserved_mem) / GPU memory capacity.
+
+      The coefficient 1.5 is a heuristic value, in the future, we can do better estimation by looking at the model types, hidden sizes or even do a dummy run.
+    """
+    cfg = resolving_view(server_args)
+    # A copy, so an earlier declaration keeps the value it recorded.
+    cuda_graph_config = copy.deepcopy(cfg.cuda_graph_config)
+    decode_cuda_graph_config = cuda_graph_config.decode
+    prefill_cuda_graph_config = cuda_graph_config.prefill
+
+    if gpu_mem is not None:
+        if gpu_mem < 20 * 1024:
+            # T4, 4080
+            # (chunked_prefill_size 2k, max_bs 8)
+            if cfg.chunked_prefill_size is None:
+                declare_resolution(
+                    server_args,
+                    "_handle_gpu_memory_settings",
+                    chunked_prefill_size=2048,
+                )
+            if decode_cuda_graph_config.max_bs is None:
+                decode_cuda_graph_config.max_bs = 8
+        elif gpu_mem < 35 * 1024:
+            # A10, 4090, 5090
+            # (chunked_prefill_size 2k, max_bs 24 if tp < 4 else 80)
+            if cfg.chunked_prefill_size is None:
+                declare_resolution(
+                    server_args,
+                    "_handle_gpu_memory_settings",
+                    chunked_prefill_size=2048,
+                )
+            if decode_cuda_graph_config.max_bs is None:
+                if cfg.tp_size < 4:
+                    decode_cuda_graph_config.max_bs = 24
+                else:
+                    decode_cuda_graph_config.max_bs = 80
+        elif gpu_mem < 60 * 1024:
+            # A100 (40GB), L40,
+            # (chunked_prefill_size 4k, max_bs 32 if tp < 4 else 160)
+            if cfg.chunked_prefill_size is None:
+                declare_resolution(
+                    server_args,
+                    "_handle_gpu_memory_settings",
+                    chunked_prefill_size=4096,
+                )
+            if decode_cuda_graph_config.max_bs is None:
+                if cfg.tp_size < 4:
+                    decode_cuda_graph_config.max_bs = 32
+                else:
+                    decode_cuda_graph_config.max_bs = 160
+        elif gpu_mem < 90 * 1024:
+            # H100, A100
+            # (chunked_prefill_size 8k, max_bs 256 if tp < 4 else 512)
+            if cfg.chunked_prefill_size is None:
+                declare_resolution(
+                    server_args,
+                    "_handle_gpu_memory_settings",
+                    chunked_prefill_size=8192,
+                )
+            if decode_cuda_graph_config.max_bs is None:
+                if cfg.tp_size < 4:
+                    decode_cuda_graph_config.max_bs = 256
+                else:
+                    decode_cuda_graph_config.max_bs = 512
+        elif gpu_mem < 160 * 1024:
+            # H20, H200
+            # (chunked_prefill_size 8k, max_bs 256 if tp < 4 else 512)
+            if cfg.chunked_prefill_size is None:
+                declare_resolution(
+                    server_args,
+                    "_handle_gpu_memory_settings",
+                    chunked_prefill_size=8192,
+                )
+            if decode_cuda_graph_config.max_bs is None:
+                if cfg.tp_size < 4:
+                    decode_cuda_graph_config.max_bs = 256
+                else:
+                    decode_cuda_graph_config.max_bs = 512
+        else:
+            # B200, MI300
+            # (chunked_prefill_size 16k, max_bs 512)
+            if cfg.chunked_prefill_size is None:
+                declare_resolution(
+                    server_args,
+                    "_handle_gpu_memory_settings",
+                    chunked_prefill_size=16384,
+                )
+            if decode_cuda_graph_config.max_bs is None:
+                decode_cuda_graph_config.max_bs = 512
+    else:
+        # Fallback defaults when gpu_mem is None
+        if cfg.chunked_prefill_size is None:
+            declare_resolution(
+                server_args,
+                "_handle_gpu_memory_settings",
+                chunked_prefill_size=4096,
+            )
+        if decode_cuda_graph_config.max_bs is None:
+            decode_cuda_graph_config.max_bs = 160
+
+    # Set cuda graph batch sizes
+    if cfg.device != "cpu":
+        if decode_cuda_graph_config.bs is None:
+            decode_cuda_graph_config.bs = (
+                server_args._generate_decode_cuda_graph_batch_sizes(
+                    decode_cuda_graph_config.max_bs
+                )
+            )
+        else:
+            decode_cuda_graph_config.max_bs = max(decode_cuda_graph_config.bs)
+    else:
+        # Reuse decode_cuda_graph_config.bs for cpu graph and use torch_compile_max_bs for cpu graph batch size limit,
+        # as cpu graph is based on torch.compile
+        if decode_cuda_graph_config.bs is not None:
+            declare_resolution(
+                server_args,
+                "_handle_gpu_memory_settings",
+                torch_compile_max_bs=max(decode_cuda_graph_config.bs),
+            )
+        else:
+            # If decode_cuda_graph_config.bs is not set, we will preferentially use torch_compile_max_bs
+            # to generate decode_cuda_graph_config.bs
+            declare_resolution(
+                server_args,
+                "_handle_gpu_memory_settings",
+                torch_compile_max_bs=cfg.torch_compile_max_bs
+                or decode_cuda_graph_config.max_bs,
+            )
+            decode_cuda_graph_config.bs = server_args._generate_cpu_graph_batch_sizes()
+
+        assert (
+            cfg.torch_compile_max_bs > 0
+        ), "cuda_graph_config[decode].bs should contain positive batch sizes"
+        decode_cuda_graph_config.max_bs = cfg.torch_compile_max_bs
+
+    if prefill_cuda_graph_config.max_bs is None:
+        # Refer to pr #15927, by default we set the prefill max_bs to the chunked prefill size.
+        # For MLA backend, the introduction of piecewise cuda graph will influence the kernel dispatch difference compared to the original mode.
+        # To avoid the performance regression, we set max_bs to 2048 by default.
+        if not server_args.use_mla_backend():
+            prefill_cuda_graph_config.max_bs = cfg.chunked_prefill_size
+        else:
+            prefill_cuda_graph_config.max_bs = 2048
+
+        # If max_total_tokens is set, cap prefill max_bs to not exceed max_total_tokens.
+        if cfg.max_total_tokens is not None:
+            prefill_cuda_graph_config.max_bs = min(
+                prefill_cuda_graph_config.max_bs, cfg.max_total_tokens
+            )
+
+        # For Llama2 series models, max_bs is limited to 4096.
+        # TODO(yuwei): remove this after the issue is fixed
+        if "llama-2" in cfg.model_path.lower():
+            prefill_cuda_graph_config.max_bs = min(
+                prefill_cuda_graph_config.max_bs, 4096
+            )
+
+    if prefill_cuda_graph_config.bs is None:
+        prefill_cuda_graph_config.bs = (
+            server_args._generate_prefill_cuda_graph_batch_sizes(
+                prefill_cuda_graph_config.max_bs
+            )
+        )
+
+    if cuda_graph_config != cfg.cuda_graph_config:
+        declare_resolution(
+            server_args,
+            "_handle_gpu_memory_settings",
+            cuda_graph_config=cuda_graph_config,
+        )
+
+    if cfg.mem_fraction_static is None:
+        if server_args.post_capture_kv_sizing_planned():
+            # Post-capture sizing measures free memory after graph capture, so
+            # skip the graph/activation reserve; keep only the floor + parallel slack.
+            reserved_mem = 1536
+            reserved_mem += cfg.tp_size * cfg.pp_size / 8 * 1024
+        else:
+            # Tokens the activation working set scales with (per serving mode).
+            if cfg.disaggregation_mode == "decode":
+                running_requests = (
+                    cfg.max_running_requests or decode_cuda_graph_config.max_bs or 1
+                )
+                draft_tokens = cfg.speculative_num_draft_tokens or 1
+                activation_tokens = max(running_requests * draft_tokens, 2048)
+            elif cfg.chunked_prefill_size > 0:
+                activation_tokens = max(cfg.chunked_prefill_size, 2048)
+            else:
+                activation_tokens = max(cfg.max_prefill_tokens, 2048)
+            # Constant meta data (e.g., from attention backend) + activation slack.
+            reserved_mem = 512
+            reserved_mem += activation_tokens * 1.5
+            # Some adjustments for large parallel size
+            reserved_mem += cfg.tp_size * cfg.pp_size / 8 * 1024
+            reserved_mem += server_args.reserve_for_graph_mb()
+            if gpu_mem is not None and gpu_mem > 60 * 1024:
+                reserved_mem = max(reserved_mem, 10 * 1024)
+            # Reserve headroom for DeepEP all-to-all buffers on top of the floor.
+            reserved_mem += server_args.reserve_for_deepep_a2a_mb()
+
+        declare_resolution(
+            server_args,
+            "_handle_gpu_memory_settings",
+            mem_fraction_static=(
+                round((gpu_mem - reserved_mem) / gpu_mem, 3)
+                if gpu_mem is not None
+                else 0.88
+            ),
+        )
+
+        # Multimodal models need more memory for the image processing,
+        # so we adjust the mem_fraction_static accordingly. The VLM encoder
+        # only runs on the prefill stage, so PD decode engines do not need
+        # this headroom; prefill engines and normal (non-PD) engines do.
+        model_config = server_args.get_model_config()
+        if (
+            model_config.is_multimodal
+            and not cfg.language_only
+            and not cfg.language_model_only
+            and cfg.disaggregation_mode != "decode"
+        ):
+            server_args.adjust_mem_fraction_for_vlm(model_config)
+
+    # If symm mem is enabled and prealloc size is not set, set it to 4GB
+    if cfg.enable_symm_mem and not envs.SGLANG_SYMM_MEM_PREALLOC_GB_SIZE.is_set():
+        envs.SGLANG_SYMM_MEM_PREALLOC_GB_SIZE.set(4)
+        logger.warning(
+            "Symmetric memory is enabled, setting symmetric memory prealloc size to 4GB as default."
+            "Use environment variable SGLANG_SYMM_MEM_PREALLOC_GB_SIZE to change the prealloc size."
+        )
