+++ python/sglang/srt/server_args.py
@@ -44,7 +44,6 @@ import json
 import logging
 import math
 import os
-import random
 import tempfile
 import uuid
 from typing import Any, Callable, Dict, List, Literal, Optional, Union
@@ -67,8 +66,6 @@ from sglang.srt.arg_groups.overrides import (
     resolved_view,
     resolving_view,
 )
-from sglang.srt.configs.embedding_model_spec import BCGPrefillPolicy
-from sglang.srt.configs.linear_attn_model_registry import get_linear_attn_spec_by_arch
 from sglang.srt.connector import ConnectorType
 from sglang.srt.distributed.device_communicators.mooncake_transfer_engine import (
     parse_ib_device_config,
@@ -91,32 +88,23 @@ from sglang.srt.speculative.decoupled_spec_io import DecoupledSpecIpcConfig
 from sglang.srt.utils.common import (
     LORA_TARGET_ALL_MODULES,
     SUPPORTED_LORA_TARGET_MODULES,
-    get_device,
     get_device_memory_capacity,
-    get_device_sm,
-    get_quantization_config,
     human_readable_int,
     is_cuda,
     is_flashinfer_available,
     is_hip,
     is_hopper_with_cuda_12_3,
-    is_mnnvl_fabric_device,
     is_mps,
     is_musa,
     is_no_spec_infer_or_topk_one,
     is_npu,
-    is_remote_url,
-    is_sm90_supported,
-    is_sm100_or_sm110_supported,
     is_sm100_supported,
-    is_sm120_supported,
     is_xpu,
     json_list_type,
     nullable_str,
     parse_connector_type,
     torch_release,
 )
-from sglang.srt.utils.hf_transformers_utils import check_gguf_file
 from sglang.srt.utils.network import NetworkAddress, get_free_port, wait_port_available
 from sglang.srt.utils.runai_utils import is_runai_obj_uri
 
@@ -4012,227 +4000,14 @@ class ServerArgs:
         self._resolution_finished = True
 
     def _handle_return_hidden_states_mode(self):
-        cfg = resolving_view(self)
-        if cfg.return_hidden_states_mode not in (None, "last", "full"):
-            raise ValueError(
-                "return_hidden_states_mode must be one of: None, 'last', or 'full'."
-            )
-        if cfg.return_hidden_states_mode is None:
-            if cfg.enable_return_hidden_states:
-                self._declare(
-                    "_handle_return_hidden_states_mode",
-                    return_hidden_states_mode="full",
-                )
-        else:
-            self._declare(
-                "_handle_return_hidden_states_mode",
-                enable_return_hidden_states=True,
-            )
+        from sglang.srt.arg_groups.serving_hook import handle_return_hidden_states_mode
+
+        handle_return_hidden_states_mode(self)
 
     def _handle_model_capability_adjustments(self):
-        cfg = resolving_view(self)
-        if parse_connector_type(cfg.model_path) == ConnectorType.INSTANCE:
-            return
-        from sglang.srt.arg_groups.overrides import (
-            _hrm_text_attention_force,
-            run_post_process_pass,
-        )
+        from sglang.srt.arg_groups.model_hook import handle_model_capability_adjustments
 
-        model_config = self.get_model_config()
-        hf_config = model_config.hf_config
-
-        # HRM-Text needs bidirectional prompt attention (prefill), which only
-        # the Triton backend honors at the kernel level. Radix/prefix reuse is
-        # also unsafe: the recurrent forward writes direction-dependent KV
-        # across many slots.
-        is_hrm_text = getattr(
-            hf_config, "model_type", None
-        ) == "hrm_text" or "HrmTextForCausalLM" in getattr(
-            hf_config, "architectures", []
-        )
-        # prefix_lm defaults to True upstream; defaulting False would skip the
-        # bidirectional-attention forcing and silently produce junk output.
-        if is_hrm_text and getattr(hf_config, "prefix_lm", True):
-            run_post_process_pass(self, _hrm_text_attention_force)
-            self._declare(
-                "_handle_model_capability_adjustments",
-                chunked_prefill_size=-1,
-            )
-            self._declare(
-                "_handle_model_capability_adjustments",
-                disable_radix_cache=True,
-            )
-            self._declare(
-                "_handle_model_capability_adjustments",
-                disable_cuda_graph=True,
-            )
-            # cuda_graph_config was already parsed from the legacy boolean, so
-            # flipping the boolean alone would not stop graph capture.
-            self._declare(
-                "_handle_model_capability_adjustments",
-                cuda_graph_config=with_phase(
-                    cfg.cuda_graph_config, Phase.DECODE, backend=Backend.DISABLED
-                ),
-            )
-            self._declare(
-                "_handle_model_capability_adjustments",
-                cuda_graph_config=with_phase(
-                    cfg.cuda_graph_config, Phase.PREFILL, backend=Backend.DISABLED
-                ),
-            )
-            logger.warning(
-                "HRM-Text (prefix_lm) detected: forcing --attention-backend "
-                "triton, --chunked-prefill-size -1, --disable-radix-cache, and "
-                "--disable-cuda-graph for correctness of the bidirectional "
-                "prompt attention."
-            )
-
-        # EmbeddingGemma is a Gemma3TextModel with bidirectional prompt
-        # attention. Prefix reuse and split prefills would reuse K/V states
-        # whose values depend on later prompt tokens, so both are invalid.
-        # Breakable CUDA Graph captures one complete prefill and is the graph
-        # mode validated for this encoder-style attention.
-        # Native encoder architectures declare a pooling-only task and do not
-        # need the legacy --is-embedding intent flag. Decoder checkpoints still
-        # require that explicit opt-in because their architecture alone does
-        # not distinguish embedding from generation serving.
-        #
-        # ``_handle_model_capability_adjustments`` is also exercised directly
-        # by a few focused tests that use a small ModelConfig stand-in. Keep
-        # the old predicate as a compatibility fallback while production
-        # ModelConfig instances use the central capability contract.
-        embedding_model_spec = getattr(model_config, "embedding_model_spec", None)
-        if (
-            embedding_model_spec is not None
-            and embedding_model_spec.auto_enable_embedding
-            and not cfg.is_embedding
-        ):
-            self._declare(
-                "_handle_model_capability_adjustments",
-                is_embedding=True,
-            )
-            logger.info(
-                "Embedding architecture detected: enabling embedding mode automatically."
-            )
-
-        is_embedding_gemma = (
-            embedding_model_spec is not None
-            and embedding_model_spec.bcg_prefill_policy == BCGPrefillPolicy.FULL_ENCODER
-        )
-        if embedding_model_spec is None:
-            is_embedding_gemma = getattr(model_config, "is_embedding_gemma", False)
-        if is_embedding_gemma:
-            # This is an encoder-only model even though its HF architecture is
-            # named Gemma3TextModel. Marking it as embedding mode enables the
-            # FlashAttention raw-K/V fast path, which does not write or read
-            # the paged KV cache during its single prefill forward.
-            self._declare(
-                "_handle_model_capability_adjustments",
-                is_embedding=True,
-            )
-            self._declare(
-                "_handle_model_capability_adjustments",
-                disable_radix_cache=True,
-            )
-            self._declare(
-                "_handle_model_capability_adjustments",
-                chunked_prefill_size=-1,
-            )
-            # Submit a list-valued embeddings request atomically so BCG can
-            # replay its full prefill batch instead of starting item zero
-            # while the remaining texts are still being tokenized.
-            self._declare(
-                "_handle_model_capability_adjustments",
-                enable_tokenizer_batch_encode=True,
-            )
-            requested_prefill_backend = (
-                cfg.prefill_attention_backend or cfg.attention_backend
-            )
-            if (
-                is_cuda()
-                and (is_sm90_supported() or is_sm100_supported())
-                and requested_prefill_backend in (None, "fa3", "fa4")
-            ):
-                # Hopper/Blackwell's default FA backend can consume raw K/V
-                # tensors for a single embedding prefill. Enable its no-KV
-                # pool path before memory-pool sizing; an explicit non-FA
-                # backend retains the existing paged-KV behavior.
-                self._declare(
-                    "_handle_model_capability_adjustments",
-                    prefill_only_disable_kv_cache=True,
-                )
-                self._validate_prefill_only_disable_kv_cache_args()
-            self._declare(
-                "_handle_model_capability_adjustments",
-                cuda_graph_config=with_phase(
-                    cfg.cuda_graph_config, Phase.DECODE, backend=Backend.DISABLED
-                ),
-            )
-            if is_cuda() and cfg.cuda_graph_config.prefill.backend != Backend.DISABLED:
-                self._declare(
-                    "_handle_model_capability_adjustments",
-                    cuda_graph_config=with_phase(
-                        cfg.cuda_graph_config, Phase.PREFILL, backend=Backend.BREAKABLE
-                    ),
-                )
-                # CUDA-graph sizing has already run by this point and derives
-                # its generic maximum from the 8K chunked-prefill default.
-                # On the Hopper/Blackwell FA raw-K/V path, raise the unlocked
-                # default to a full eight-way 2K embedding batch; callers can
-                # still override this for larger aggregate prefills.
-                prefill_config = cfg.cuda_graph_config.prefill
-                # Unit-level capability tests may invoke this hook without
-                # running the full CUDA-graph configuration parser, which is
-                # where this internal lock set is normally initialized.
-                # Treat that minimal construction as having no user-locked
-                # graph settings.
-                cuda_graph_config_locked = getattr(
-                    self, "_cuda_graph_config_locked", set()
-                )
-                if (Phase.PREFILL, "max_bs") not in cuda_graph_config_locked:
-                    sizing = {
-                        "max_bs": max(
-                            prefill_config.max_bs or 0,
-                            model_config.context_len,
-                            16384,
-                        )
-                    }
-                    if (Phase.PREFILL, "bs") not in cuda_graph_config_locked:
-                        sizing["bs"] = self._generate_prefill_cuda_graph_batch_sizes(
-                            sizing["max_bs"]
-                        )
-                    self._declare(
-                        "_handle_model_capability_adjustments",
-                        cuda_graph_config=with_phase(
-                            cfg.cuda_graph_config, Phase.PREFILL, **sizing
-                        ),
-                    )
-            elif not is_cuda():
-                # BCG is CUDA-only. Other graph backends do not support this
-                # encoder-style prefill, so retain the eager Triton path.
-                self._declare(
-                    "_handle_model_capability_adjustments",
-                    cuda_graph_config=with_phase(
-                        cfg.cuda_graph_config, Phase.PREFILL, backend=Backend.DISABLED
-                    ),
-                )
-            logger.info(
-                "EmbeddingGemma detected: disabling radix cache and chunked "
-                "prefill; using breakable CUDA graph for CUDA prefill."
-            )
-
-        if (
-            model_config.is_multimodal
-            and not model_config.is_multimodal_chunked_prefill_supported
-        ):
-            self._declare(
-                "_handle_model_capability_adjustments",
-                chunked_prefill_size=-1,
-            )
-            logger.info(
-                f"Automatically turn off --chunked-prefill-size as it is not supported for "
-                f"{hf_config.model_type}"
-            )
+        handle_model_capability_adjustments(self)
 
     def _handle_model_source_paths(self):
         from sglang.srt.arg_groups.model_path_hook import handle_model_source_paths
@@ -4272,235 +4047,19 @@ class ServerArgs:
         handle_media_url_security(self)
 
     def _handle_deprecated_args(self):
-        cfg = resolving_view(self)
-        if cfg.disable_fast_image_processor:
-            if cfg.image_processor_backend not in {"auto", "pil"}:
-                raise ValueError(
-                    "--disable-fast-image-processor conflicts with "
-                    f"--image-processor-backend={cfg.image_processor_backend}."
-                )
-            logger.warning(
-                "--disable-fast-image-processor is deprecated; use "
-                "--image-processor-backend=pil instead."
-            )
-            self._declare("_handle_deprecated_args", image_processor_backend="pil")
+        from sglang.srt.arg_groups.serving_hook import handle_deprecated_args
 
-        # Handle deprecated tool call parsers
-        deprecated_tool_call_parsers = {"qwen25": "qwen", "glm45": "glm"}
-        if cfg.tool_call_parser in deprecated_tool_call_parsers:
-            logger.warning(
-                f"The tool_call_parser '{cfg.tool_call_parser}' is deprecated. Please use '{deprecated_tool_call_parsers[cfg.tool_call_parser]}' instead."
-            )
-            self._declare(
-                "_handle_deprecated_args",
-                tool_call_parser=deprecated_tool_call_parsers[cfg.tool_call_parser],
-            )
-
-        # When user passes --enable-flashinfer-allreduce-fusion, enable with auto backend
-        if (
-            cfg.enable_flashinfer_allreduce_fusion
-            and cfg.flashinfer_allreduce_fusion_backend is None
-        ):
-            logger.warning(
-                "--enable-flashinfer-allreduce-fusion is deprecated. "
-                "Please use --flashinfer-allreduce-fusion-backend=auto instead."
-            )
-            self._declare(
-                "_handle_deprecated_args",
-                flashinfer_allreduce_fusion_backend="auto",
-            )
-        self._declare(
-            "_handle_deprecated_args",
-            enable_flashinfer_allreduce_fusion=False,
-        )
-        # Deprecated attention-backend alias: "compressed" -> "dsv4".
-        renamed = {}
-        for attr in (
-            "attention_backend",
-            "decode_attention_backend",
-            "prefill_attention_backend",
-            "speculative_draft_attention_backend",
-        ):
-            if getattr(self, attr, None) == "compressed":
-                logger.warning(
-                    "--%s=compressed is deprecated; use 'dsv4' instead.",
-                    attr.replace("_", "-"),
-                )
-                renamed[attr] = "dsv4"
-        if renamed:
-            self._declare("_handle_deprecated_args", **renamed)
-
-        # --grpc-mode is a deprecated alias for --smg-grpc-mode.
-        if cfg.grpc_mode and not cfg.smg_grpc_mode:
-            logger.warning(
-                "--grpc-mode is deprecated and will be removed in a future "
-                "version. Use --smg-grpc-mode for the legacy SMG gRPC server, "
-                "or --grpc-port for the native gRPC server."
-            )
-            self._declare(
-                "_handle_deprecated_args",
-                smg_grpc_mode=True,
-            )
-
-        # Native gRPC tuning knob is env-only; --grpc-port (CLI) enables the
-        # native server, falling back to SGLANG_GRPC_PORT.
-        self._declare(
-            "_handle_deprecated_args",
-            grpc_worker_threads=envs.SGLANG_GRPC_WORKER_THREADS.get(),
-        )
-
-        grpc_port_env = envs.SGLANG_GRPC_PORT.get()
-        if cfg.grpc_port is None and grpc_port_env is not None:
-            self._declare(
-                "_handle_deprecated_args",
-                grpc_port=grpc_port_env,
-            )
-
-        # Legacy SMG defaults its port to --port + 10000. Derive/validate only
-        # when gRPC is in use, so HTTP-only high ports don't fail validation.
-        legacy_grpc = cfg.smg_grpc_mode or cfg.grpc_mode
-        if legacy_grpc and cfg.grpc_port is None:
-            self._declare(
-                "_handle_deprecated_args",
-                grpc_port=cfg.port + 10000,
-            )
-
-        if cfg.grpc_port is not None:
-            if not (1 <= cfg.grpc_port <= 65535):
-                raise ValueError(
-                    "--grpc-port / SGLANG_GRPC_PORT "
-                    f"({cfg.grpc_port}) must be between 1 and 65535"
-                )
-            if cfg.grpc_worker_threads is not None and cfg.grpc_worker_threads < 1:
-                raise ValueError(
-                    "SGLANG_GRPC_WORKER_THREADS "
-                    f"({cfg.grpc_worker_threads}) must be >= 1"
-                )
-
-        # Native gRPC is incompatible with launch paths it doesn't wire into.
-        # Legacy takes precedence over grpc_port, keeping re-runs idempotent.
-        native_grpc = cfg.grpc_port is not None and not legacy_grpc
-        if cfg.sidecar_args is not None:
-            if cfg.sidecar is None:
-                raise ValueError("--sidecar-args requires --sidecar.")
-            if not isinstance(cfg.sidecar_args, list) or not all(
-                isinstance(arg, str) for arg in cfg.sidecar_args
-            ):
-                raise ValueError("--sidecar-args must be a JSON array of strings.")
-        if cfg.sidecar is not None:
-            if not cfg.sidecar.strip():
-                raise ValueError("--sidecar must not be empty.")
-            if legacy_grpc:
-                raise ValueError(
-                    "--sidecar requires SGLang's native gRPC server; "
-                    "it cannot be combined with --smg-grpc-mode/--grpc-mode."
-                )
-            if cfg.grpc_port is None:
-                raise ValueError("--sidecar requires --grpc-port or SGLANG_GRPC_PORT.")
-        if native_grpc:
-            if cfg.use_ray:
-                raise ValueError(
-                    "--grpc-port is not supported with --use-ray: the Ray "
-                    "serve launch path does not start the native gRPC server."
-                )
-            if cfg.encoder_only:
-                raise ValueError(
-                    "--grpc-port is not supported with --encoder-only: "
-                    "encoder disaggregation uses its own server."
-                )
-            if cfg.tokenizer_worker_num > 1:
-                raise ValueError(
-                    "Native gRPC does not yet support --tokenizer-worker-num > 1. "
-                    "Unset --grpc-port or set --tokenizer-worker-num 1."
-                )
-            if cfg.api_key or cfg.admin_api_key:
-                raise ValueError(
-                    "--grpc-port is incompatible with --api-key/--admin-api-key: "
-                    "the native gRPC listener bypasses HTTP auth middleware."
-                )
+        handle_deprecated_args(self)
 
     def _handle_prefill_delayer_env_compat(self):
-        if envs.SGLANG_SCHEDULER_DECREASE_PREFILL_IDLE.get():
-            self._declare(
-                "_handle_prefill_delayer_env_compat",
-                enable_prefill_delayer=True,
-            )
-        if x := envs.SGLANG_PREFILL_DELAYER_MAX_DELAY_PASSES.get():
-            self._declare(
-                "_handle_prefill_delayer_env_compat",
-                prefill_delayer_max_delay_passes=x,
-            )
-        if x := envs.SGLANG_PREFILL_DELAYER_TOKEN_USAGE_LOW_WATERMARK.get():
-            self._declare(
-                "_handle_prefill_delayer_env_compat",
-                prefill_delayer_token_usage_low_watermark=x,
-            )
+        from sglang.srt.arg_groups.serving_hook import handle_prefill_delayer_env_compat
+
+        handle_prefill_delayer_env_compat(self)
 
     def _handle_missing_default_values(self):
-        cfg = resolving_view(self)
-        if cfg.tokenizer_path is None:
-            self._declare(
-                "_handle_missing_default_values",
-                tokenizer_path=cfg.model_path,
-            )
-        if cfg.served_model_name is None:
-            self._declare(
-                "_handle_missing_default_values",
-                served_model_name=cfg.model_path,
-            )
-        if cfg.device is None:
-            self._declare(
-                "_handle_missing_default_values",
-                device=get_device(),
-            )
-        # strip device index from user if any (e.g. "cuda:0" -> "cuda")
-        self._declare(
-            "_handle_missing_default_values",
-            device=cfg.device.split(":")[0],
-        )
-        if cfg.random_seed is None:
-            self._declare(
-                "_handle_missing_default_values",
-                random_seed=random.randint(0, 1 << 30),
-            )
-        if cfg.mm_process_config is None:
-            self._declare("_handle_missing_default_values", mm_process_config={})
+        from sglang.srt.arg_groups.serving_hook import handle_missing_default_values
 
-        # Handle ModelScope model downloads
-        if envs.SGLANG_USE_MODELSCOPE.get():
-            self._handle_modelscope_paths()
-
-        # In speculative scenario:
-        # - If `speculative_draft_model_quantization` is specified, the draft model uses this quantization method.
-        # - Otherwise, the draft model defaults to the same quantization as the target model.
-        if cfg._speculative_draft_quantization_explicitly_set is None:
-            self._declare(
-                "_handle_missing_default_values",
-                _speculative_draft_quantization_explicitly_set=cfg.speculative_draft_model_quantization
-                is not None,
-            )
-        if cfg.speculative_draft_model_quantization is None:
-            self._declare(
-                "_handle_missing_default_values",
-                speculative_draft_model_quantization=cfg.quantization,
-            )
-
-        # Resolve --quantization unquant before model config validation. Record
-        # the explicit opt-out so later auto-detection does not re-enable
-        # quantization.
-        if cfg.quantization == "unquant":
-            self._declare(
-                "_handle_missing_default_values",
-                quantization=None,
-            )
-            self._quantization_explicitly_unset = True
-        else:
-            self._quantization_explicitly_unset = False
-        if cfg.speculative_draft_model_quantization == "unquant":
-            self._declare(
-                "_handle_missing_default_values",
-                speculative_draft_model_quantization=None,
-            )
+        handle_missing_default_values(self)
 
     def _handle_modelscope_paths(self):
         from sglang.srt.arg_groups.model_path_hook import handle_modelscope_paths
@@ -4641,243 +4200,9 @@ class ServerArgs:
         handle_multi_item_scoring(self)
 
     def _handle_gpu_memory_settings(self, gpu_mem):
-        """
-        Configure GPU memory-dependent settings including
-        chunked_prefill_size, cuda_graph_config[decode].max_bs, and mem_fraction_static.
+        from sglang.srt.arg_groups.memory_hook import handle_gpu_memory_settings
 
-        Here are our heuristics:
-        - Set chunked_prefill_size and cuda_graph_config[decode].max_bs based on the GPU memory capacity.
-          This is because GPUs with more memory are generally more powerful, we need to use a larger
-          chunked_prefill_size and a larger decode max_bs to fully utilize the GPU.
-        - Then set mem_fraction_static based on chunked_prefill_size and decode max_bs.
-
-          GPU memory capacity = model weights + KV cache pool + activations + cuda graph buffers
-
-          The argument mem_fraction_static is defined as (model weights + KV cache pool) / GPU memory capacity,
-          or equivalently, mem_fraction_static = (GPU memory capacity - activations - cuda graph buffers) / GPU memory capacity.
-
-          In order to compute mem_fraction_static, we need to estimate the size of activations and cuda graph buffers.
-          The activation memory is proportional to the chunked_prefill_size.
-          The cuda graph memory is proportional to the decode max_bs.
-          We use reserved_mem = chunked_prefill_size * 1.5 + max_bs * 2 to estimate the size of activations and cuda graph buffers in GB,
-          and set mem_fraction_static = (GPU memory capacity - reserved_mem) / GPU memory capacity.
-
-          The coefficient 1.5 is a heuristic value, in the future, we can do better estimation by looking at the model types, hidden sizes or even do a dummy run.
-        """
-        cfg = resolving_view(self)
-        # A copy, so an earlier declaration keeps the value it recorded.
-        cuda_graph_config = copy.deepcopy(cfg.cuda_graph_config)
-        decode_cuda_graph_config = cuda_graph_config.decode
-        prefill_cuda_graph_config = cuda_graph_config.prefill
-
-        if gpu_mem is not None:
-            if gpu_mem < 20 * 1024:
-                # T4, 4080
-                # (chunked_prefill_size 2k, max_bs 8)
-                if cfg.chunked_prefill_size is None:
-                    self._declare(
-                        "_handle_gpu_memory_settings",
-                        chunked_prefill_size=2048,
-                    )
-                if decode_cuda_graph_config.max_bs is None:
-                    decode_cuda_graph_config.max_bs = 8
-            elif gpu_mem < 35 * 1024:
-                # A10, 4090, 5090
-                # (chunked_prefill_size 2k, max_bs 24 if tp < 4 else 80)
-                if cfg.chunked_prefill_size is None:
-                    self._declare(
-                        "_handle_gpu_memory_settings",
-                        chunked_prefill_size=2048,
-                    )
-                if decode_cuda_graph_config.max_bs is None:
-                    if cfg.tp_size < 4:
-                        decode_cuda_graph_config.max_bs = 24
-                    else:
-                        decode_cuda_graph_config.max_bs = 80
-            elif gpu_mem < 60 * 1024:
-                # A100 (40GB), L40,
-                # (chunked_prefill_size 4k, max_bs 32 if tp < 4 else 160)
-                if cfg.chunked_prefill_size is None:
-                    self._declare(
-                        "_handle_gpu_memory_settings",
-                        chunked_prefill_size=4096,
-                    )
-                if decode_cuda_graph_config.max_bs is None:
-                    if cfg.tp_size < 4:
-                        decode_cuda_graph_config.max_bs = 32
-                    else:
-                        decode_cuda_graph_config.max_bs = 160
-            elif gpu_mem < 90 * 1024:
-                # H100, A100
-                # (chunked_prefill_size 8k, max_bs 256 if tp < 4 else 512)
-                if cfg.chunked_prefill_size is None:
-                    self._declare(
-                        "_handle_gpu_memory_settings",
-                        chunked_prefill_size=8192,
-                    )
-                if decode_cuda_graph_config.max_bs is None:
-                    if cfg.tp_size < 4:
-                        decode_cuda_graph_config.max_bs = 256
-                    else:
-                        decode_cuda_graph_config.max_bs = 512
-            elif gpu_mem < 160 * 1024:
-                # H20, H200
-                # (chunked_prefill_size 8k, max_bs 256 if tp < 4 else 512)
-                if cfg.chunked_prefill_size is None:
-                    self._declare(
-                        "_handle_gpu_memory_settings",
-                        chunked_prefill_size=8192,
-                    )
-                if decode_cuda_graph_config.max_bs is None:
-                    if cfg.tp_size < 4:
-                        decode_cuda_graph_config.max_bs = 256
-                    else:
-                        decode_cuda_graph_config.max_bs = 512
-            else:
-                # B200, MI300
-                # (chunked_prefill_size 16k, max_bs 512)
-                if cfg.chunked_prefill_size is None:
-                    self._declare(
-                        "_handle_gpu_memory_settings",
-                        chunked_prefill_size=16384,
-                    )
-                if decode_cuda_graph_config.max_bs is None:
-                    decode_cuda_graph_config.max_bs = 512
-        else:
-            # Fallback defaults when gpu_mem is None
-            if cfg.chunked_prefill_size is None:
-                self._declare(
-                    "_handle_gpu_memory_settings",
-                    chunked_prefill_size=4096,
-                )
-            if decode_cuda_graph_config.max_bs is None:
-                decode_cuda_graph_config.max_bs = 160
-
-        # Set cuda graph batch sizes
-        if cfg.device != "cpu":
-            if decode_cuda_graph_config.bs is None:
-                decode_cuda_graph_config.bs = (
-                    self._generate_decode_cuda_graph_batch_sizes(
-                        decode_cuda_graph_config.max_bs
-                    )
-                )
-            else:
-                decode_cuda_graph_config.max_bs = max(decode_cuda_graph_config.bs)
-        else:
-            # Reuse decode_cuda_graph_config.bs for cpu graph and use torch_compile_max_bs for cpu graph batch size limit,
-            # as cpu graph is based on torch.compile
-            if decode_cuda_graph_config.bs is not None:
-                self._declare(
-                    "_handle_gpu_memory_settings",
-                    torch_compile_max_bs=max(decode_cuda_graph_config.bs),
-                )
-            else:
-                # If decode_cuda_graph_config.bs is not set, we will preferentially use torch_compile_max_bs
-                # to generate decode_cuda_graph_config.bs
-                self._declare(
-                    "_handle_gpu_memory_settings",
-                    torch_compile_max_bs=cfg.torch_compile_max_bs
-                    or decode_cuda_graph_config.max_bs,
-                )
-                decode_cuda_graph_config.bs = self._generate_cpu_graph_batch_sizes()
-
-            assert (
-                cfg.torch_compile_max_bs > 0
-            ), "cuda_graph_config[decode].bs should contain positive batch sizes"
-            decode_cuda_graph_config.max_bs = cfg.torch_compile_max_bs
-
-        if prefill_cuda_graph_config.max_bs is None:
-            # Refer to pr #15927, by default we set the prefill max_bs to the chunked prefill size.
-            # For MLA backend, the introduction of piecewise cuda graph will influence the kernel dispatch difference compared to the original mode.
-            # To avoid the performance regression, we set max_bs to 2048 by default.
-            if not self.use_mla_backend():
-                prefill_cuda_graph_config.max_bs = cfg.chunked_prefill_size
-            else:
-                prefill_cuda_graph_config.max_bs = 2048
-
-            # If max_total_tokens is set, cap prefill max_bs to not exceed max_total_tokens.
-            if cfg.max_total_tokens is not None:
-                prefill_cuda_graph_config.max_bs = min(
-                    prefill_cuda_graph_config.max_bs, cfg.max_total_tokens
-                )
-
-            # For Llama2 series models, max_bs is limited to 4096.
-            # TODO(yuwei): remove this after the issue is fixed
-            if "llama-2" in cfg.model_path.lower():
-                prefill_cuda_graph_config.max_bs = min(
-                    prefill_cuda_graph_config.max_bs, 4096
-                )
-
-        if prefill_cuda_graph_config.bs is None:
-            prefill_cuda_graph_config.bs = (
-                self._generate_prefill_cuda_graph_batch_sizes(
-                    prefill_cuda_graph_config.max_bs
-                )
-            )
-
-        if cuda_graph_config != cfg.cuda_graph_config:
-            self._declare(
-                "_handle_gpu_memory_settings", cuda_graph_config=cuda_graph_config
-            )
-
-        if cfg.mem_fraction_static is None:
-            if self.post_capture_kv_sizing_planned():
-                # Post-capture sizing measures free memory after graph capture, so
-                # skip the graph/activation reserve; keep only the floor + parallel slack.
-                reserved_mem = 1536
-                reserved_mem += cfg.tp_size * cfg.pp_size / 8 * 1024
-            else:
-                # Tokens the activation working set scales with (per serving mode).
-                if cfg.disaggregation_mode == "decode":
-                    running_requests = (
-                        cfg.max_running_requests or decode_cuda_graph_config.max_bs or 1
-                    )
-                    draft_tokens = cfg.speculative_num_draft_tokens or 1
-                    activation_tokens = max(running_requests * draft_tokens, 2048)
-                elif cfg.chunked_prefill_size > 0:
-                    activation_tokens = max(cfg.chunked_prefill_size, 2048)
-                else:
-                    activation_tokens = max(cfg.max_prefill_tokens, 2048)
-                # Constant meta data (e.g., from attention backend) + activation slack.
-                reserved_mem = 512
-                reserved_mem += activation_tokens * 1.5
-                # Some adjustments for large parallel size
-                reserved_mem += cfg.tp_size * cfg.pp_size / 8 * 1024
-                reserved_mem += self.reserve_for_graph_mb()
-                if gpu_mem is not None and gpu_mem > 60 * 1024:
-                    reserved_mem = max(reserved_mem, 10 * 1024)
-                # Reserve headroom for DeepEP all-to-all buffers on top of the floor.
-                reserved_mem += self.reserve_for_deepep_a2a_mb()
-
-            self._declare(
-                "_handle_gpu_memory_settings",
-                mem_fraction_static=(
-                    round((gpu_mem - reserved_mem) / gpu_mem, 3)
-                    if gpu_mem is not None
-                    else 0.88
-                ),
-            )
-
-            # Multimodal models need more memory for the image processing,
-            # so we adjust the mem_fraction_static accordingly. The VLM encoder
-            # only runs on the prefill stage, so PD decode engines do not need
-            # this headroom; prefill engines and normal (non-PD) engines do.
-            model_config = self.get_model_config()
-            if (
-                model_config.is_multimodal
-                and not cfg.language_only
-                and not cfg.language_model_only
-                and cfg.disaggregation_mode != "decode"
-            ):
-                self.adjust_mem_fraction_for_vlm(model_config)
-
-        # If symm mem is enabled and prealloc size is not set, set it to 4GB
-        if cfg.enable_symm_mem and not envs.SGLANG_SYMM_MEM_PREALLOC_GB_SIZE.is_set():
-            envs.SGLANG_SYMM_MEM_PREALLOC_GB_SIZE.set(4)
-            logger.warning(
-                "Symmetric memory is enabled, setting symmetric memory prealloc size to 4GB as default."
-                "Use environment variable SGLANG_SYMM_MEM_PREALLOC_GB_SIZE to change the prealloc size."
-            )
+        handle_gpu_memory_settings(self, gpu_mem)
 
     def post_capture_kv_sizing_planned(self) -> bool:
         """Whether the mem_fraction heuristic may skip the graph reserve; must be
@@ -5115,559 +4440,9 @@ class ServerArgs:
         validate_hisparse_kv_cache_dtype(self)
 
     def _handle_model_specific_adjustments(self):
-        cfg = resolving_view(self)
-        from sglang.srt.configs.model_config import (
-            get_mimo_v2_fused_qkv_expected_tp_size,
-            is_deepseek_dsa,
-        )
+        from sglang.srt.arg_groups.model_hook import handle_model_specific_adjustments
 
-        if cfg.enable_deterministic_inference:
-            self._declare(
-                "_handle_model_specific_adjustments",
-                enforce_disable_flashinfer_allreduce_fusion=True,
-            )
-
-        self._declare(
-            "_handle_model_specific_adjustments",
-            uses_mamba_radix_cache=False,
-        )
-        if parse_connector_type(cfg.model_path) == ConnectorType.INSTANCE:
-            # No model overrides for an instance connector: no hf_config to
-            # key them on.
-            return
-
-        model_config = self.get_model_config()
-        hf_config = model_config.hf_config
-        model_arch = hf_config.architectures[0]
-
-        if model_arch == "InternS2MobiusForConditionalGeneration":
-            unsupported = []
-            if cfg.pp_size != 1:
-                unsupported.append("pipeline parallelism (--pp-size must be 1)")
-            if cfg.ep_size != 1:
-                unsupported.append("expert parallelism (--ep-size must be 1)")
-            if unsupported:
-                raise ValueError(
-                    "Intern-S2-Mobius does not support: " + "; ".join(unsupported) + "."
-                )
-
-        if cfg.enable_dsa_cache_layer_split and not is_deepseek_dsa(hf_config):
-            raise ValueError(
-                "--enable-dsa-cache-layer-split is only supported for DSA "
-                "(DeepSeek Sparse Attention) models."
-            )
-
-        if cfg.enable_cp_decode_attn_tp:
-            from sglang.srt.layers.cp.cp_decode_attn_tp import (
-                CP_DECODE_ATTN_TP_SUPPORTED_ARCHS,
-            )
-
-            if model_arch not in CP_DECODE_ATTN_TP_SUPPORTED_ARCHS:
-                raise ValueError(
-                    "--enable-cp-decode-attn-tp is only supported for models "
-                    "whose attention linears are replicated across CP ranks "
-                    f"(attn_tp_size=1). Got {model_arch}; supported: "
-                    f"{sorted(CP_DECODE_ATTN_TP_SUPPORTED_ARCHS)}."
-                )
-
-        _hybrid_spec = get_linear_attn_spec_by_arch(model_arch)
-        if _hybrid_spec is not None and _hybrid_spec.uses_mamba_radix_cache:
-            self._handle_mamba_radix_cache(model_arch=model_arch)
-
-        # Collect the declarative model overrides (registry) on the
-        # pristine config and stash them for publish-time flags resolution;
-        # server_args is never mutated — mid-resolution readers see the
-        # declared values through resolved_view, runtime readers through the
-        # flags tier.
-        from sglang.srt.arg_groups.overrides import (
-            collect_model_override_declarations,
-            validate_declarations,
-        )
-
-        model_overrides = collect_model_override_declarations(
-            model_arch, self, hf_config
-        )
-        validate_declarations(self, model_overrides)
-        self._resolved_overrides.extend(model_overrides)
-
-        if model_arch in (
-            "KimiLinearForCausalLM",
-            "KimiK3ForConditionalGeneration",
-        ):
-            from sglang.srt.arg_groups.kimi_k3_hook import (
-                apply_kimi_k3_linear_attn_defaults,
-                apply_kimi_k3_spec_backend_defaults,
-            )
-
-            apply_kimi_k3_linear_attn_defaults(self)
-            apply_kimi_k3_spec_backend_defaults(self)
-
-        if model_arch in [
-            "DeepseekV4ForCausalLM",
-        ]:
-            from sglang.srt.arg_groups.deepseek_v4_hook import (
-                apply_deepseek_v4_defaults,
-            )
-
-            apply_deepseek_v4_defaults(self, model_arch)
-
-        if model_arch in [
-            "DeepseekV3ForCausalLM",
-            "DeepseekV32ForCausalLM",
-            "KimiK25ForConditionalGeneration",
-            "MistralLarge3ForCausalLM",
-            "PixtralForConditionalGeneration",
-            "GlmMoeDsaForCausalLM",
-            "LongcatFlashForCausalLM",
-            "Dots3NoteForCausalLM",
-        ]:
-            # Set attention backend for DeepSeek
-            if is_deepseek_dsa(hf_config):  # DeepSeek 3.2/GLM 5
-                if envs.SGLANG_DSA_PREFILL_DENSE_ATTN_KV_LEN_THRESHOLD.is_set():
-                    logger.warning(
-                        f"Dense attention kv len threshold is manually set to {envs.SGLANG_DSA_PREFILL_DENSE_ATTN_KV_LEN_THRESHOLD.get()} for DSA. Caution: This may cause performance regression if the threshold is larger than the index topk of model."
-                    )
-                else:
-                    # When threshold is not manually set, set it to the index topk of model
-                    from sglang.srt.configs.model_config import get_dsa_index_topk
-
-                    envs.SGLANG_DSA_PREFILL_DENSE_ATTN_KV_LEN_THRESHOLD.set(
-                        get_dsa_index_topk(hf_config)
-                    )
-                    logger.warning(
-                        f"Set dense attention kv len threshold to model index_topk={envs.SGLANG_DSA_PREFILL_DENSE_ATTN_KV_LEN_THRESHOLD.get()} for DeepSeek with DSA."
-                    )
-                # The "dsa" attention fill moved to the override registry
-                # (arg_groups/overrides.py: _deepseek_family_overrides).
-
-                index_topk_freq = getattr(hf_config, "index_topk_freq", 1) or 1
-                index_topk_pattern = getattr(hf_config, "index_topk_pattern", None)
-                if cfg.enable_two_batch_overlap and (
-                    index_topk_freq > 1
-                    or (index_topk_pattern is not None and "S" in index_topk_pattern)
-                ):
-                    raise ValueError(
-                        "--enable-two-batch-overlap is not supported with DSA "
-                        "index-topk sharing (index_topk_freq > 1 or an "
-                        "index_topk_pattern containing shared layers): the TBO op "
-                        "path does not propagate topk indices across layers, so "
-                        "shared layers would run sparse attention without indices."
-                    )
-
-                if not is_npu() and not is_xpu():  # CUDA or ROCm GPU
-                    if cfg.enable_prefill_cp:
-                        # The DSA CP field declarations moved to the override
-                        # registry (arg_groups/overrides.py:
-                        # _deepseek_family_overrides).
-                        self._declare(
-                            "_handle_model_specific_adjustments",
-                            cuda_graph_config=with_phase(
-                                cfg.cuda_graph_config,
-                                Phase.PREFILL,
-                                backend=Backend.DISABLED,
-                            ),
-                        )
-                    else:
-                        # Pure TP and partial DP Attention mode is active for DSA, logging a warning
-                        if cfg.dp_size < cfg.tp_size:
-                            logger.warning(
-                                f"DSA with TP mode is active, dp_size={cfg.dp_size}, tp_size={cfg.tp_size}, "
-                                f"attn_tp_size={cfg.tp_size}, attention weights will be sharded across {cfg.tp_size} ranks."
-                            )
-
-                    # The DSA page-size selection moved to the override registry
-                    # (arg_groups/overrides.py: _deepseek_family_overrides).
-
-                    import torch
-
-                    major, _ = torch.cuda.get_device_capability()
-                    self._set_default_dsa_kv_cache_dtype(
-                        major, resolved_view(self).quantization
-                    )
-                    self._set_default_dsa_backends(major)
-
-                if cfg.enable_prefill_cp:
-                    assert (
-                        cfg.disaggregation_mode != "decode"
-                    ), "CP is only supported for prefill when PD disaggregation, please remove --enable-prefill-cp."
-                if (
-                    cfg.enable_dsa_cache_layer_split
-                    and cfg.disaggregation_mode != "prefill"
-                ):
-                    if cfg.disaggregation_mode == "decode":
-                        raise ValueError(
-                            "--enable-dsa-cache-layer-split is not supported on "
-                            "decode workers. This flag is a prefill-CP "
-                            "optimization; decode receives full cache shards "
-                            "through PD transfer."
-                        )
-                    raise ValueError(
-                        "--enable-dsa-cache-layer-split is only supported on PD "
-                        "prefill workers. Non-PD workers also run decode and "
-                        "require ordinary local decode cache semantics."
-                    )
-                if cfg.enable_dsa_cache_layer_split and (
-                    not cfg.enable_prefill_cp or cfg.cp_strategy != "interleave"
-                ):
-                    raise ValueError(
-                        "--enable-dsa-cache-layer-split requires "
-                        "--enable-prefill-cp and --cp-strategy interleave "
-                        "(or legacy --enable-nsa-prefill-context-parallel with "
-                        "--nsa-prefill-cp-mode round-robin-split)."
-                    )
-                # Layer split relies on the mooncake all-CP-rank KV/indexer
-                # transfer path. mori/nixl support is a temporary limitation
-                # and will be added later by the community.
-                if (
-                    cfg.enable_dsa_cache_layer_split
-                    and cfg.disaggregation_transfer_backend != "mooncake"
-                ):
-                    raise ValueError(
-                        "--enable-dsa-cache-layer-split currently only supports "
-                        "the mooncake transfer backend (mooncake / mooncake_tcp). "
-                        f"Got --disaggregation-transfer-backend "
-                        f"{cfg.disaggregation_transfer_backend!r}. mori/nixl "
-                        "support will be added later by the community."
-                    )
-                if cfg.enable_dsa_cache_layer_split and cfg.pp_size > 1:
-                    raise ValueError(
-                        "--enable-dsa-cache-layer-split is not supported with "
-                        "pipeline parallelism (pp_size > 1) yet. It requires "
-                        "prefill context parallelism, and CP + PP has not been "
-                        "validated for this feature."
-                    )
-
-            else:
-                # DeepSeek V3/R1/V3.1
-                if cfg.cuda_graph_config.prefill.backend != Backend.DISABLED:
-                    logger.info("Piecewise CUDA graph is enabled, use MLA for prefill.")
-
-                # The sm100 trtllm_mla fill moved to the override registry
-                # (arg_groups/overrides.py: _deepseek_family_overrides).
-
-                # MLA prefill CP auto-config: the field declarations moved to
-                # the override registry (arg_groups/overrides.py:
-                # _deepseek_family_overrides).
-                if cfg.enable_prefill_cp and self.use_mla_backend():
-                    self._declare(
-                        "_handle_model_specific_adjustments",
-                        cuda_graph_config=with_phase(
-                            cfg.cuda_graph_config,
-                            Phase.PREFILL,
-                            backend=Backend.DISABLED,
-                        ),
-                    )
-
-            # Set moe backend for DeepSeek: the sm100 quant/moe resolution
-            # moved to the resolution pipeline (arg_groups/overrides.py:
-            # _deepseek_moe_quant_resolution -- a slot pass, because the DSA
-            # kv-cache-dtype default above must read the pristine
-            # quantization). The HIP arm (fusion log + spec_moe writes, the
-            # latter awaiting the speculative-hook migration) stays below.
-            from sglang.srt.arg_groups.overrides import (
-                _deepseek_moe_quant_resolution,
-                run_post_process_pass,
-            )
-
-            run_post_process_pass(self, _deepseek_moe_quant_resolution)
-            if is_hip():
-                if is_deepseek_dsa(hf_config):
-                    # The fused top-k v2 kernel (topk_transform_512_v2) is a
-                    # CUDA/Hopper-only path: its JIT source includes
-                    # <cooperative_groups.h> and uses cg::this_cluster()
-                    # (thread-block clusters), neither of which exists on ROCm,
-                    # so it fails to JIT-compile on gfx9xx during CUDA-graph
-                    # capture. DeepSeek-V4 already disables it on HIP; mirror that
-                    # here for the rest of the DSA family (DeepSeek-V3.2 /
-                    # GLM-5.x) that shares the same decode top-k path.
-                    envs.SGLANG_OPT_USE_TOPK_V2.set(False)
-                if not self._resolved().enable_dp_attention and cfg.nnodes == 1:
-                    # TODO (Hubert): Put this back later
-                    # self.enable_aiter_allreduce_fusion = True
-                    logger.info(
-                        "Enable Aiter AllReduce Fusion for DeepseekV3ForCausalLM"
-                    )
-
-                # The fp4-checkpoint draft spec-MoE resolution moved to the
-                # resolution pipeline (arg_groups/overrides.py:
-                # _deepseek_spec_moe_resolution), invoked here at its legacy
-                # slot.
-                from sglang.srt.arg_groups.overrides import (
-                    _deepseek_spec_moe_resolution,
-                )
-
-                run_post_process_pass(self, _deepseek_spec_moe_resolution)
-
-        elif model_arch in [
-            "DeepseekV4ForCausalLM",
-        ]:
-            from sglang.srt.arg_groups.deepseek_v4_hook import (
-                validate_deepseek_v4_cp,
-                validate_deepseek_v4_mega_moe_token_budget,
-            )
-
-            validate_deepseek_v4_cp(self)
-            validate_deepseek_v4_mega_moe_token_budget(self)
-
-            if is_sm120_supported():
-                # SM120 lacks tcgen05/TMEM: disable features that depend on
-                # DeepGEMM or require >99KB SMEM (topk_v2).
-                envs.SGLANG_OPT_FP8_WO_A_GEMM.set(False)
-                envs.SGLANG_OPT_USE_TOPK_V2.set(False)
-                envs.SGLANG_OPT_USE_TILELANG_MHC_PRE.set(False)
-                if not envs.SGLANG_OPT_FUSE_MHC_POST_PRE.is_set():
-                    envs.SGLANG_OPT_FUSE_MHC_POST_PRE.set(True)
-                envs.SGLANG_OPT_DEEPGEMM_HC_PRENORM.set(False)
-                envs.SGLANG_FP8_PAGED_MQA_LOGITS_TORCH.set(True)
-                # Prefer TileLang over the Torch fallback.
-                envs.SGLANG_OPT_USE_TILELANG_INDEXER.set(True)
-            elif is_hip():
-                envs.SGLANG_OPT_DEEPGEMM_HC_PRENORM.set(False)
-                envs.SGLANG_OPT_FP8_WO_A_GEMM.set(False)
-                envs.SGLANG_OPT_USE_JIT_INDEXER_METADATA.set(False)
-                envs.SGLANG_OPT_USE_TOPK_V2.set(True)
-                envs.SGLANG_OPT_USE_AITER_INDEXER.set(True)
-                envs.SGLANG_OPT_USE_TILELANG_MHC_PRE.set(False)
-                envs.SGLANG_OPT_USE_TILELANG_MHC_POST.set(False)
-                envs.SGLANG_FP8_PAGED_MQA_LOGITS_TORCH.set(True)
-                envs.SGLANG_OPT_USE_MULTI_STREAM_OVERLAP.set(False)
-                envs.SGLANG_EAGER_INPUT_NO_COPY.set(True)
-
-        elif model_arch in ["GptOssForCausalLM"]:
-            # Attention backend selection + XPU dtype validation moved to the
-            # override registry (arg_groups/overrides.py: _gpt_oss_overrides).
-            # Exempt MLX only: none of these backends exist on MPS, and MLX runs
-            # attention inside its own runner, so attention_backend is still
-            # unset here.  Plain macOS stays on the list -- torch_native has
-            # neither sliding window nor attention sinks.
-            if not (is_mps() and use_mlx()):
-                supported_backends = [
-                    "triton",
-                    "trtllm_mha",
-                    "fa3",
-                    "fa4",
-                    "ascend",
-                    "intel_amx",
-                    "intel_xpu",
-                    "aiter",
-                ]
-                prefill_attn_backend, decode_attn_backend = (
-                    self._resolved_attention_backends()
-                )
-                assert (
-                    prefill_attn_backend in supported_backends
-                    and decode_attn_backend in supported_backends
-                ), (
-                    f"GptOssForCausalLM requires one of {supported_backends} attention backend, but got the following backends\n"
-                    f"- Prefill: {prefill_attn_backend}\n"
-                    f"- Decode: {decode_attn_backend}\n"
-                )
-
-            quant_method = get_quantization_config(hf_config)
-            is_mxfp4_quant_format = quant_method == "mxfp4"
-            if (
-                not self._resolved().enable_dp_attention
-                and cfg.nnodes == 1
-                and is_hip()
-            ):
-                # TODO (Hubert): Put this back later
-                # self.enable_aiter_allreduce_fusion = True
-                logger.info("Enable Aiter AllReduce Fusion for GptOssForCausalLM")
-            quantization_config = getattr(hf_config, "quantization_config", None)
-            is_mxfp4_quant_format = (
-                quantization_config is not None
-                and quantization_config.get("quant_method") == "mxfp4"
-            )
-            # The mxfp4 dtype override moved to the override registry
-            # (arg_groups/overrides.py: _gpt_oss_overrides).
-
-            # The moe_runner_backend selection moved to the override registry
-            # (arg_groups/overrides.py: _gpt_oss_overrides).
-
-            if resolved_view(self).moe_runner_backend == "triton_kernel":
-                assert (
-                    self._resolved().ep_size == 1
-                ), "Triton kernel MoE is only supported when ep_size == 1"
-
-        elif model_arch in ("MiMoV2ForCausalLM", "MiMoV2FlashForCausalLM"):
-            if model_arch == "MiMoV2ForCausalLM" and not cfg.encoder_only:
-                expected_attn_tp_size = get_mimo_v2_fused_qkv_expected_tp_size(
-                    hf_config
-                )
-                view = self._resolved()
-                attn_dp_size = cfg.dp_size if view.enable_dp_attention else 1
-                effective_attn_tp_size = (
-                    cfg.tp_size // attn_dp_size // view.attn_cp_size
-                )
-                if (
-                    expected_attn_tp_size is not None
-                    and expected_attn_tp_size % effective_attn_tp_size != 0
-                ):
-                    raise ValueError(
-                        "MiMoV2ForCausalLM requires effective attention TP "
-                        f"size {expected_attn_tp_size} because its fused "
-                        "qkv_proj weights are "
-                        f"TP={expected_attn_tp_size}-interleaved; got "
-                        f"{effective_attn_tp_size} "
-                        f"(tp_size={cfg.tp_size}, dp_size={cfg.dp_size}, "
-                        f"enable_dp_attention={view.enable_dp_attention}, "
-                        f"attn_cp_size={view.attn_cp_size}). "
-                        "Set --tp, --dp, --enable-dp-attention, and "
-                        "--attention-context-parallel-size so the effective "
-                        f"attention TP size is {expected_attn_tp_size}."
-                    )
-
-            # enable_multi_layer_eagle for EAGLE moved to the override registry
-            # (arg_groups/overrides.py: _mimo_v2_overrides).
-
-            # MiMoV2 hierarchical cache runs on the unified radix tree, which
-            # is the default tree cache now. MiMoV2 has head_dim != v_head_dim,
-            # so the host KV pool uses asymmetric K/V allocation. Both
-            # kernel/page_first and direct/page_first_direct have split K/V
-            # transfer paths.
-        elif (
-            "Step3p5ForCausalLM" in model_arch
-            or "Step3p7ForConditionalGeneration" in model_arch
-        ):
-            # Attention backend selection + EAGLE multi-layer +
-            # hierarchical-cache SWA writes moved to the override registry
-            # (arg_groups/overrides.py: _step3p_overrides).
-            pass
-        elif (
-            model_arch in ("Llama4ForConditionalGeneration", "Llama4ForCausalLM")
-            and cfg.device != "cpu"
-        ):
-            # Attention backend auto-select moved to the override registry
-            # (arg_groups/overrides.py: _llama4_overrides).
-            attention_backend = resolved_view(self).attention_backend
-            assert attention_backend in {
-                "fa3",
-                "aiter",
-                "triton",
-                "ascend",
-                "trtllm_mha",
-                "intel_xpu",
-            }, f"fa3, aiter, triton, ascend, trtllm_mha or intel_xpu is required for Llama4 model but got {attention_backend}"
-            # The moe_runner_backend selection moved to the override registry
-            # (arg_groups/overrides.py: _llama4_overrides).
-        # Gemma2/Gemma3 (disable_hybrid_swa_memory) moved to the override registry
-        # (arg_groups/overrides.py: _gemma2_gemma3_overrides).
-        elif model_arch in (
-            "Gemma4ForConditionalGeneration",
-            "Gemma4ForCausalLM",
-            "Gemma4UnifiedForConditionalGeneration",
-        ):
-            # Default attention backend selection moved to the override registry
-            # (arg_groups/overrides.py: _gemma4_overrides).
-            prefill_backend, decode_backend = self._resolved_attention_backends()
-            accepted_backends = (
-                "trtllm_mha",
-                "triton",
-                "ascend",
-                "intel_xpu",
-                "intel_amx",
-            )
-            assert (
-                prefill_backend in accepted_backends
-                and decode_backend in accepted_backends
-            ), (
-                "Gemma4 only supports trtllm_mha, triton, ascend, intel_xpu, or intel_amx "
-                f"attention backend, got prefill={prefill_backend}, decode={decode_backend}"
-            )
-
-            # The quantization/moe_runner_backend resolution moved to the override
-            # registry (arg_groups/overrides.py: _gemma4_overrides).
-        elif model_arch == "MossVLForConditionalGeneration":
-            # The prefill attention backend default + validation moved to the
-            # override registry (arg_groups/overrides.py: _moss_vl_overrides).
-            pass
-        elif model_arch in ["Exaone4ForCausalLM", "ExaoneMoEForCausalLM"]:
-            if hf_config.sliding_window_pattern is not None:
-                # disable_hybrid_swa_memory moved to the override registry
-                # (arg_groups/overrides.py: _exaone_overrides).
-                # https://docs.sglang.ai/advanced_features/attention_backend.html
-                accepted_backends = ["fa3", "triton", "trtllm_mha"]
-                attention_backend = resolved_view(self).attention_backend
-                assert (
-                    attention_backend in accepted_backends
-                ), f"One of the attention backends in {accepted_backends} is required for {model_arch}, but got {attention_backend}"
-        elif model_arch in ["Olmo2ForCausalLM"]:
-            # disable_hybrid_swa_memory + attention backend selection moved to
-            # the override registry (arg_groups/overrides.py: _olmo2_overrides).
-
-            # Flashinfer appears to degrade performance when sliding window attention
-            # is used for the Olmo2 architecture. Olmo2 does not use sliding window attention
-            # but Olmo3 does.
-            attention_backend = resolved_view(self).attention_backend
-            assert (
-                attention_backend != "flashinfer"
-            ), "FlashInfer backend can significantly degrade the performance of Olmo3 models."
-
-            logger.info(
-                f"Using {attention_backend} as attention backend for {model_arch}."
-            )
-        elif model_arch in [
-            "Qwen3MoeForCausalLM",
-            "Qwen3VLMoeForConditionalGeneration",
-            "Qwen3NextForCausalLM",
-            "Qwen3_5MoeForConditionalGeneration",
-            "InternS2PreviewForConditionalGeneration",
-            "Qwen3_5ForConditionalGeneration",
-        ]:
-            # The quantization/moe_runner_backend resolution moved to the
-            # override registry (arg_groups/overrides.py:
-            # _qwen3_moe_family_overrides); the hybrid sub-family's attention
-            # backend + page size defaults to _qwen3_5_hybrid_overrides.
-            pass
-
-        elif model_arch in ["Glm4MoeForCausalLM"]:
-            # The quantization/moe_runner_backend/enable_tf32_matmul resolution
-            # moved to the override registry (arg_groups/overrides.py:
-            # _glm4_moe_overrides).
-            pass
-
-        elif model_arch in ["Lfm2ForCausalLM", "Lfm2MoeForCausalLM"]:
-            # Attention backend selection moved to the override registry
-            # (arg_groups/overrides.py: _lfm2_overrides).
-            assert resolved_view(self).attention_backend != "triton", (
-                f"{model_arch} does not support triton attention backend, "
-                "as the first layer might not be an attention layer"
-            )
-
-        # MiniMaxM2ForCausalLM (enable_tf32_matmul) moved to the override registry
-        # (arg_groups/overrides.py: _minimax_m2_overrides).
-
-        # Qwen3VL aiter unified-attention page_size moved to the override registry
-        # (arg_groups/overrides.py: _qwen3vl_overrides).
-
-        # Hybrid-mamba radix cache handling for the per-arch branch call sites
-        # dissolved above: the resolution pass self-guards on the arch union
-        # (and the Granite layer_types probe), so one call covers them all.
-        # Hybrid-spec archs already resolved at the pre-dispatch call above;
-        # for them this re-invocation is an idempotent no-op plus validation.
-        # Kept ahead of the sparse-head pass: the legacy per-branch calls
-        # resolved before that tail write of disable_overlap_schedule.
-        self._handle_mamba_radix_cache(model_arch=model_arch)
-
-        from sglang.srt.arg_groups.overrides import (
-            _sparse_head_overlap_disable,
-            run_post_process_pass,
-        )
-
-        run_post_process_pass(self, _sparse_head_overlap_disable)
-
-        # The FlashInfer AllReduce Fusion auto-enable and the enforce-disable
-        # terminal moved to the resolution pipeline (arg_groups/overrides.py:
-        # _flashinfer_allreduce_fusion_auto_enable /
-        # _enforce_disable_allreduce_fusion), invoked here at their legacy
-        # slots.
-        from sglang.srt.arg_groups.overrides import (
-            _enforce_disable_allreduce_fusion,
-            _flashinfer_allreduce_fusion_auto_enable,
-        )
-
-        run_post_process_pass(self, _flashinfer_allreduce_fusion_auto_enable)
-        run_post_process_pass(self, _enforce_disable_allreduce_fusion)
+        handle_model_specific_adjustments(self)
 
     def _support_mamba_cache_extra_buffer(self, model_arch: str):
         from sglang.srt.arg_groups.overrides import supports_mamba_cache_extra_buffer
@@ -5724,21 +4499,9 @@ class ServerArgs:
         # Resolution moved to the resolution pipeline (arg_groups/overrides.py:
         # _mamba_radix_cache_resolution), invoked here at each legacy call
         # slot; this handler keeps the validation.
-        from sglang.srt.arg_groups.overrides import (
-            _mamba_radix_cache_resolution,
-            mamba_extra_buffer_of,
-            run_post_process_pass,
-        )
+        from sglang.srt.arg_groups.model_hook import handle_mamba_radix_cache
 
-        run_post_process_pass(self, _mamba_radix_cache_resolution)
-        view = resolved_view(self)
-        if not view.uses_mamba_radix_cache:
-            return
-
-        if mamba_extra_buffer_of(view):
-            self._validate_mamba_extra_buffer(view, model_arch)
-        else:
-            self._validate_mamba_no_buffer(view, model_arch)
+        handle_mamba_radix_cache(self, model_arch)
 
     def _handle_sampling_backend(self):
         # Moved to the resolution pipeline (arg_groups/overrides.py:
@@ -6262,120 +5025,9 @@ class ServerArgs:
         # The quantization side of the gguf coupling moved to the pipeline
         # (arg_groups/overrides.py: _gguf_quantization); load_format itself is
         # genuine config (runtime user updates write it) and stays imperative.
-        cfg = resolving_view(self)
-        from sglang.srt.arg_groups.overrides import (
-            _gguf_quantization,
-            run_post_process_pass,
-        )
+        from sglang.srt.arg_groups.model_path_hook import handle_load_format
 
-        run_post_process_pass(self, _gguf_quantization)
-        if (cfg.load_format == "auto" or cfg.load_format == "gguf") and check_gguf_file(
-            cfg.model_path
-        ):
-            self._declare(
-                "_handle_load_format",
-                load_format="gguf",
-            )
-
-        if cfg.load_format == "auto" and self._is_mistral_native_format():
-            self._declare(
-                "_handle_load_format",
-                load_format="mistral",
-            )
-            logger.info(
-                "Detected Mistral native format checkpoint, setting load_format='mistral'"
-            )
-
-        if is_runai_obj_uri(cfg.model_path):
-            self._declare(
-                "_handle_load_format",
-                load_format="runai_streamer",
-            )
-        elif is_remote_url(cfg.model_path):
-            self._declare(
-                "_handle_load_format",
-                load_format="remote",
-            )
-
-        if (
-            cfg.speculative_draft_model_path is not None
-            and is_runai_obj_uri(cfg.speculative_draft_model_path)
-            and cfg.speculative_draft_load_format is None
-        ):
-            self._declare(
-                "_handle_load_format",
-                speculative_draft_load_format="runai_streamer",
-            )
-
-        if cfg.custom_weight_loader is None:
-            self._declare("_handle_load_format", custom_weight_loader=[])
-
-        if cfg.load_format == "remote_instance":
-            if cfg.remote_instance_weight_loader_backend != "modelexpress" and (
-                cfg.remote_instance_weight_loader_seed_instance_ip is None
-                or cfg.remote_instance_weight_loader_seed_instance_service_port is None
-            ):
-                logger.warning(
-                    "Fallback load_format to 'auto' due to incomplete remote instance weight loader settings."
-                )
-                self._declare(
-                    "_handle_load_format",
-                    load_format="auto",
-                )
-            elif (
-                cfg.remote_instance_weight_loader_send_weights_group_ports is None
-                and cfg.remote_instance_weight_loader_backend == "nccl"
-            ):
-                logger.warning(
-                    "Fallback load_format to 'auto' due to incomplete remote instance weight loader NCCL group ports settings."
-                )
-                self._declare(
-                    "_handle_load_format",
-                    load_format="auto",
-                )
-            elif (
-                cfg.remote_instance_weight_loader_backend == "transfer_engine"
-                and not self.validate_transfer_engine()
-            ):
-                logger.warning(
-                    "Fallback load_format to 'auto' due to 'transfer_engine' backend is not supported."
-                )
-                self._declare(
-                    "_handle_load_format",
-                    load_format="auto",
-                )
-
-        # Check whether TransferEngine can be used when users want to start seed service that supports TransferEngine backend.
-        if cfg.remote_instance_weight_loader_start_seed_via_transfer_engine:
-            self._declare(
-                "_handle_load_format",
-                remote_instance_weight_loader_start_seed_via_transfer_engine=self.validate_transfer_engine(),
-            )
-
-        # "ipc_cache" is an internal-only load format: ModelRunner sets it
-        # automatically when the weight cache is enabled, and it is not a public
-        # --load-format choice. Setting it directly is always wrong (no daemon is
-        # launched, and fallback_load_format inherits a nonsensical format), so
-        # reject it and point at the knob (defense-in-depth; the CLI already
-        # rejects it via LOAD_FORMAT_CHOICES).
-        if cfg.load_format == "ipc_cache":
-            raise ValueError(
-                "load_format='ipc_cache' is an internal-only format and must not "
-                "be set directly. Enable the weight cache via --weight-cache-mode "
-                "client (connect to an existing daemon) or daemon (launch one); "
-                "that selects IPC loading automatically."
-            )
-
-        # Speculative decoding loads an extra draft model whose weights the
-        # daemon does not export, so refuse the combination up front instead of
-        # failing deep inside draft-worker load (draft-model daemon TBD).
-        if cfg.weight_cache_mode != "off" and cfg.speculative_algorithm is not None:
-            raise ValueError(
-                "--weight-cache-mode is not supported together with speculative "
-                "decoding (--speculative-algorithm): the weight cache daemon does "
-                "not export the draft model's weights. Disable one of them "
-                "(--weight-cache-mode off) for this configuration."
-            )
+        handle_load_format(self)
 
     def _is_mistral_native_format(self) -> bool:
         """True iff the checkpoint requires load_format=mistral.
@@ -6440,107 +5092,16 @@ class ServerArgs:
     LANGUAGE_MODEL_ONLY_ARCHITECTURES = ("MuseGlimmerForConditionalGeneration",)
 
     def _handle_language_model_only(self):
-        cfg = resolving_view(self)
-        if not cfg.language_model_only:
-            return
-        for flag, name in (
-            (cfg.encoder_only, "--encoder-only"),
-            (cfg.language_only, "--language-only"),
-            (cfg.enable_prefix_mm_cache, "--enable-prefix-mm-cache"),
-            (
-                cfg.enable_broadcast_mm_inputs_process,
-                "--enable-broadcast-mm-inputs-process",
-            ),
-            (cfg.mm_enable_dp_encoder, "--mm-enable-dp-encoder"),
-        ):
-            if flag:
-                raise ValueError(
-                    f"--language-model-only cannot be combined with {name}"
-                )
-        if cfg.disaggregation_mode != "null":
-            raise ValueError(
-                "--language-model-only is incompatible with --disaggregation-mode "
-                "prefill/decode"
-            )
-        architectures = self.get_model_config().hf_config.architectures
-        if not any(a in self.LANGUAGE_MODEL_ONLY_ARCHITECTURES for a in architectures):
-            raise ValueError(
-                f"--language-model-only does not support {architectures}. "
-                f"Supported: {list(self.LANGUAGE_MODEL_ONLY_ARCHITECTURES)}."
-            )
+        from sglang.srt.arg_groups.model_hook import handle_language_model_only
+
+        handle_language_model_only(self)
 
     def _handle_encoder_disaggregation(self):
-        cfg = resolving_view(self)
-        self._handle_language_model_only()
-        if cfg.enable_prefix_mm_cache and not cfg.encoder_only:
-            raise ValueError(
-                "--enable-prefix-mm-cache requires --encoder-only to be enabled"
-            )
-        if cfg.encoder_only and cfg.language_only:
-            raise ValueError("Cannot set --encoder-only and --language-only together")
-        if cfg.encoder_only and not cfg.disaggregation_mode == "null":
-            raise ValueError(
-                "Cannot set --encoder-only and --disaggregation-mode prefill/decode together"
-            )
+        from sglang.srt.arg_groups.pd_disaggregation_hook import (
+            handle_encoder_disaggregation,
+        )
 
-        if cfg.language_only and len(cfg.encoder_urls) == 0:
-            logger.info(
-                "--language-only is set without --encoder-urls. Encoders are "
-                "expected to register dynamically via the "
-                "EncoderBootstrapServer."
-            )
-
-        # Validate IB devices when mooncake backend is used
-        if (
-            cfg.disaggregation_transfer_backend == "mooncake"
-            and cfg.disaggregation_mode in ("prefill", "decode")
-        ) or cfg.encoder_transfer_backend == "mooncake":
-            self._declare(
-                "_handle_encoder_disaggregation",
-                disaggregation_ib_device=self._validate_ib_devices(
-                    cfg.disaggregation_ib_device
-                ),
-            )
-
-        # Validate model type for encoder disaggregation
-        hf_config = self.get_model_config().hf_config
-        model_arch = hf_config.architectures[0]
-        if cfg.encoder_transfer_backend == "auto":
-            self._declare(
-                "_handle_encoder_disaggregation",
-                encoder_transfer_backend=resolve_encoder_transfer_backend(
-                    cfg.encoder_transfer_backend, model_arch, cfg.tp_size
-                ),
-            )
-            if cfg.encoder_only or cfg.language_only:
-                logger.info(
-                    "Encoder transfer backend auto-resolved to %s for %s at TP%d.",
-                    cfg.encoder_transfer_backend,
-                    model_arch,
-                    cfg.tp_size,
-                )
-        if (cfg.encoder_only or cfg.language_only) and model_arch not in [
-            "Qwen2VLForConditionalGeneration",
-            "Qwen3VLForConditionalGeneration",
-            "Qwen2_5_VLForConditionalGeneration",
-            "Qwen3VLMoeForConditionalGeneration",
-            "Qwen3_5ForConditionalGeneration",
-            "Qwen3_5MoeForConditionalGeneration",
-            "InternS2PreviewForConditionalGeneration",
-            "Qwen3OmniMoeForConditionalGeneration",
-            "Qwen2AudioForConditionalGeneration",
-            "Qwen2_5OmniForConditionalGeneration",
-            "Dots3NoteForCausalLM",
-            "KimiVLForConditionalGeneration",
-            "KimiK25ForConditionalGeneration",
-            "KimiK3ForConditionalGeneration",
-            "MiMoV2ForCausalLM",
-        ]:
-            raise ValueError(
-                f"Model type {model_arch} is not supported for encoder disaggregation. "
-                f"Supported architectures: Qwen2VL, Qwen3VL, Qwen3.5, InternS2, "
-                f"Qwen2Audio, Qwen2.5Omni, Dots3-Note, Kimi, MiMoV2."
-            )
+        handle_encoder_disaggregation(self)
 
     def _validate_ib_devices(self, device_str: Optional[str]) -> Optional[str]:
         """
@@ -6618,278 +5179,21 @@ class ServerArgs:
         return json.dumps(normalized_mapping, separators=(",", ":"))
 
     def _handle_tokenizer_batching(self):
-        cfg = resolving_view(self)
-        if cfg.enable_tokenizer_batch_encode and cfg.enable_dynamic_batch_tokenizer:
-            raise ValueError(
-                "Cannot enable both --enable-tokenizer-batch-encode and --enable-dynamic-batch-tokenizer. "
-                "Please choose one tokenizer batching approach."
-            )
+        from sglang.srt.arg_groups.serving_hook import handle_tokenizer_batching
 
-        if cfg.skip_tokenizer_init and not envs.SGLANG_RUST_SERVER.get():
-            # Tokenizer workers still serve HTTP / state / output work, so
-            # their fanout is preserved; detokenizer workers only decode.
-            if cfg.detokenizer_worker_num != 1:
-                logger.warning(
-                    "skip_tokenizer_init=True leaves no decode work for detokenizer workers; "
-                    f"forcing detokenizer_worker_num=1 (requested {cfg.detokenizer_worker_num})."
-                )
-                self._declare("_handle_tokenizer_batching", detokenizer_worker_num=1)
-
-            if cfg.enable_tokenizer_batch_encode:
-                logger.warning(
-                    "skip_tokenizer_init=True ignores --enable-tokenizer-batch-encode; disabling it."
-                )
-                self._declare(
-                    "_handle_tokenizer_batching",
-                    enable_tokenizer_batch_encode=False,
-                )
-
-            if cfg.enable_dynamic_batch_tokenizer:
-                logger.warning(
-                    "skip_tokenizer_init=True ignores --enable-dynamic-batch-tokenizer; disabling it."
-                )
-                self._declare(
-                    "_handle_tokenizer_batching",
-                    enable_dynamic_batch_tokenizer=False,
-                )
-
-            logger.info(
-                "skip_tokenizer_init=True: string-based stop conditions (stop, stop_regex) "
-                "and min_new_tokens are unavailable."
-            )
+        handle_tokenizer_batching(self)
 
     def _handle_multimodal_feature_transport(self):
-        """Resolve multimodal feature transport before tokenizer workers start.
-
-        CUDA IPC is opt-in because its fixed pool on ``base_gpu_id`` reduces the
-        memory left for model/KV-cache allocations. Multi-node MNNVL deployments
-        may still auto-select CUDA VMM. The legacy CUDA IPC flag and environment
-        variable remain supported so existing deployments map to this policy.
-        """
-        cfg = resolving_view(self)
-        requested_transport = cfg.mm_feature_transport
-        legacy_ipc_is_set = envs.SGLANG_USE_CUDA_IPC_TRANSPORT.is_set()
-        legacy_ipc_enabled = envs.SGLANG_USE_CUDA_IPC_TRANSPORT.get()
-
-        if cfg.keep_mm_feature_on_device:
-            if requested_transport not in (None, "cuda_ipc"):
-                raise ValueError(
-                    "--keep-mm-feature-on-device conflicts with "
-                    f"--mm-feature-transport={requested_transport}. Use only "
-                    "--mm-feature-transport=cuda_ipc."
-                )
-            requested_transport = "cuda_ipc"
-            logger.warning(
-                "--keep-mm-feature-on-device is deprecated; using "
-                "--mm-feature-transport=cuda_ipc instead."
-            )
-
-        if requested_transport is None:
-            if legacy_ipc_is_set:
-                requested_transport = "cuda_ipc" if legacy_ipc_enabled else "cpu"
-                logger.warning(
-                    "SGLANG_USE_CUDA_IPC_TRANSPORT is deprecated; use "
-                    "--mm-feature-transport=%s instead.",
-                    requested_transport,
-                )
-            elif cfg.encoder_only:
-                requested_transport = "cpu"
-                logger.info(
-                    "Multimodal feature transport auto-resolved to cpu for "
-                    "encoder-only serving; encoder outputs use "
-                    "--encoder-transfer-backend instead."
-                )
-            elif (
-                self.get_model_config().is_multimodal
-                and is_cuda()
-                and cfg.disaggregation_mode == "null"
-            ):
-                # A full GPU pool always degrades to CPU transport per tensor.
-                # Keep CUDA IPC opt-in because even an idle pool consumes HBM
-                # that would otherwise back the KV cache. Multi-node
-                # auto-selection is limited to GB200/GB300 systems where the
-                # runtime already enables the MNNVL/IMEX communication stack.
-                if cfg.nnodes == 1:
-                    requested_transport = "cpu"
-                elif is_mnnvl_fabric_device() and os.path.exists(
-                    "/dev/nvidia-caps-imex-channels/channel0"
-                ):
-                    from sglang.srt.model_loader.utils import (
-                        supports_cuda_vmm_feature_transport,
-                    )
-
-                    if supports_cuda_vmm_feature_transport(self.get_model_config()):
-                        requested_transport = "cuda_vmm"
-                        logger.info(
-                            "Multimodal feature transport auto-resolved to "
-                            "cuda_vmm (multi-node GB200/GB300 MNNVL). Pass "
-                            "--mm-feature-transport=cpu to opt out."
-                        )
-                    else:
-                        requested_transport = "cpu"
-                        logger.info(
-                            "Multimodal feature transport auto-resolved to cpu: "
-                            "the model has not opted into CUDA VMM transport."
-                        )
-                else:
-                    requested_transport = "cpu"
-                    if is_mnnvl_fabric_device():
-                        logger.info(
-                            "Multimodal feature transport auto-resolved to cpu: "
-                            "GB200/GB300 was detected but no IMEX channel is "
-                            "mounted. Configure the MNNVL compute domain or pass "
-                            "--mm-feature-transport=cuda_vmm after doing so."
-                        )
-            else:
-                requested_transport = "cpu"
-        elif legacy_ipc_is_set and legacy_ipc_enabled != (
-            requested_transport == "cuda_ipc"
-        ):
-            logger.warning(
-                "--mm-feature-transport=%s overrides the conflicting legacy "
-                "SGLANG_USE_CUDA_IPC_TRANSPORT=%s setting.",
-                requested_transport,
-                int(legacy_ipc_enabled),
-            )
-
-        if cfg.encoder_only and requested_transport in ("cuda_ipc", "cuda_vmm"):
-            logger.warning(
-                "--mm-feature-transport=%s does not control encoder-only "
-                "output transfer; using cpu for this inactive transport. Select "
-                "--encoder-transfer-backend for encoder outputs.",
-                requested_transport,
-            )
-            requested_transport = "cpu"
-
-        if requested_transport == "cuda_vmm":
-            if not is_cuda():
-                raise ValueError(
-                    "--mm-feature-transport=cuda_vmm requires NVIDIA CUDA."
-                )
-            if cfg.pp_size != 1:
-                raise ValueError(
-                    "--mm-feature-transport=cuda_vmm does not support pipeline "
-                    "parallelism."
-                )
-            if envs.SGLANG_RUST_SERVER.get():
-                raise ValueError(
-                    "--mm-feature-transport=cuda_vmm is not supported with "
-                    "SGLANG_RUST_SERVER."
-                )
-            pool_budget_mb = envs.SGLANG_MM_FEATURE_CACHE_MB.get()
-            handle_kind = "CUDA FABRIC" if cfg.nnodes > 1 else "POSIX FD"
-            logger.info(
-                "Using CUDA VMM for multimodal features with %s sharing: "
-                "reserving up to %d MiB on base GPU %d across %d tokenizer "
-                "worker(s). This reduces KV cache headroom; a full pool falls "
-                "back to inline CPU transport.",
-                handle_kind,
-                pool_budget_mb,
-                cfg.base_gpu_id,
-                cfg.tokenizer_worker_num,
-            )
-
-        if requested_transport == "cuda_ipc":
-            if not is_cuda():
-                raise ValueError(
-                    "--mm-feature-transport=cuda_ipc requires NVIDIA CUDA."
-                )
-            if cfg.nnodes != 1:
-                raise ValueError(
-                    "--mm-feature-transport=cuda_ipc only supports a single node."
-                )
-
-            pool_budget_mb = envs.SGLANG_MM_FEATURE_CACHE_MB.get()
-            logger.info(
-                "Using CUDA IPC for multimodal features: reserving up to %d MiB "
-                "on base GPU %d across %d tokenizer worker(s). This reduces KV "
-                "cache headroom; a full pool falls back to CPU transport.",
-                pool_budget_mb,
-                cfg.base_gpu_id,
-                cfg.tokenizer_worker_num,
-            )
-            logger.info(
-                "CUDA IPC pool-handle caching is %s. It reuses mappings to the "
-                "existing bounded pool without reserving another pool; set "
-                "SGLANG_USE_IPC_POOL_HANDLE_CACHE=0 to disable it.",
-                (
-                    "enabled"
-                    if envs.SGLANG_USE_IPC_POOL_HANDLE_CACHE.get()
-                    else "disabled"
-                ),
-            )
-
-        self._declare(
-            "_handle_multimodal_feature_transport",
-            mm_feature_transport=requested_transport,
+        from sglang.srt.arg_groups.serving_hook import (
+            handle_multimodal_feature_transport,
         )
-        # The bounded IPC pool owns device residency. Do not retain unpooled
-        # tensors after a pool miss, which would make HBM use request-dependent.
-        self._declare(
-            "_handle_multimodal_feature_transport",
-            keep_mm_feature_on_device=False,
-        )
-        envs.SGLANG_USE_CUDA_IPC_TRANSPORT.set(
-            "1" if requested_transport == "cuda_ipc" else "0"
-        )
+
+        handle_multimodal_feature_transport(self)
 
     def _handle_environment_variables(self):
-        cfg = resolving_view(self)
-        self._handle_multimodal_feature_transport()
-        envs.SGLANG_ENABLE_TORCH_COMPILE.set("1" if cfg.enable_torch_compile else "0")
-        if cfg.mamba_ssm_dtype is not None:
-            envs.SGLANG_MAMBA_SSM_DTYPE.set(cfg.mamba_ssm_dtype)
-        envs.SGLANG_DISABLE_OUTLINES_DISK_CACHE.set(
-            "1" if cfg.disable_outlines_disk_cache else "0"
-        )
-        envs.SGLANG_ENABLE_DETERMINISTIC_INFERENCE.set(
-            "1" if cfg.enable_deterministic_inference else "0"
-        )
-        if cfg.enable_deterministic_inference:
-            envs.SGLANG_FLASHINFER_MOE_FUSED_FINALIZE.set("0")
-        if cfg.debug_cuda_graph:
-            if not (is_cuda() or is_hip()):
-                logger.warning(
-                    "--debug-cuda-graph is not supported on non CUDA/HIP devices. "
-                    "Disabling breakable CUDA graph."
-                )
-                self._declare("_handle_environment_variables", debug_cuda_graph=False)
-            else:
-                envs.SGLANG_USE_BREAKABLE_CUDA_GRAPH.set("1")
-                logger.warning(
-                    "Debug mode for CUDA graph is enabled via breakable CUDA graph. "
-                    "All operations will run eagerly through the graph capture/replay path."
-                )
-        if cfg.enable_deepseek_v4_fp4_indexer and not (
-            is_sm100_supported() or is_sm120_supported()
-        ):
-            raise ValueError(
-                "--enable-deepseek-v4-fp4-indexer requires SM100 or SM120 GPUs with "
-                "DeepGEMM FP4 indexer support."
-            )
-        # FP8 W_o GEMM needs DeepGEMM JIT. Enable exactly where the runtime can run
-        # it, mirroring the forward scale split: the ue8m0 path
-        # (DEEPGEMM_SCALE_UE8M0, true sm100, default on) or an sm90 opt-in
-        # fp32-scale path (use FP4 expert ckpt). Disable in every other case.
-        if is_cuda() and envs.SGLANG_OPT_FP8_WO_A_GEMM.get():
-            from sglang.srt.layers import deep_gemm_wrapper
+        from sglang.srt.arg_groups.serving_hook import handle_environment_variables
 
-            sm = get_device_sm()
-            explicit = envs.SGLANG_OPT_FP8_WO_A_GEMM.is_set()
-            supported = deep_gemm_wrapper.DEEPGEMM_SCALE_UE8M0 or (
-                deep_gemm_wrapper.ENABLE_JIT_DEEPGEMM
-                and is_sm90_supported()
-                and explicit
-            )
-            if not supported and explicit:
-                logger.warning(
-                    "Disabling SGLANG_OPT_FP8_WO_A_GEMM: requires DeepGEMM JIT "
-                    "and sm100+ (Blackwell), or explicit opt-in on sm90; "
-                    "detected sm%d.",
-                    sm,
-                )
-            if not supported:
-                envs.SGLANG_OPT_FP8_WO_A_GEMM.set(False)
+        handle_environment_variables(self)
 
     def _handle_cache_compatibility(self):
         from sglang.srt.arg_groups.kv_cache_hook import handle_cache_compatibility
@@ -6897,132 +5201,9 @@ class ServerArgs:
         handle_cache_compatibility(self)
 
     def _handle_deterministic_inference(self):
-        cfg = resolving_view(self)
-        if cfg.rl_on_policy_target is not None:
-            logger.warning(
-                "Enable deterministic inference because of rl_on_policy_target."
-            )
-            self._declare(
-                "_handle_deterministic_inference",
-                enable_deterministic_inference=True,
-            )
+        from sglang.srt.arg_groups.attention_hook import handle_deterministic_inference
 
-            # For VLM
-            envs.SGLANG_VLM_CACHE_SIZE_MB.set(0)
-            # TODO remove this environment variable as a whole
-            envs.SGLANG_ENABLE_DETERMINISTIC_INFERENCE.set(True)
-
-        if cfg.enable_deterministic_inference:
-            if cfg.enable_aiter_allreduce_fusion:
-                logger.warning(
-                    "Disable --enable-aiter-allreduce-fusion because deterministic inference is enabled."
-                )
-                self._declare(
-                    "_handle_deterministic_inference",
-                    enable_aiter_allreduce_fusion=False,
-                )
-
-            # Moved to the resolution pipeline (arg_groups/overrides.py:
-            # _deterministic_allreduce_fusion_disable), invoked here at its
-            # legacy slot.
-            from sglang.srt.arg_groups.overrides import (
-                _deterministic_allreduce_fusion_disable,
-                run_post_process_pass,
-            )
-
-            run_post_process_pass(self, _deterministic_allreduce_fusion_disable)
-
-            # The forced-pytorch sampling write and the attention backend
-            # fill/validation moved to the resolution pipeline
-            # (arg_groups/overrides.py), invoked at their legacy slots.
-            from sglang.srt.arg_groups.overrides import (
-                _deterministic_attention_backend,
-                _deterministic_sampling_backend,
-                run_post_process_pass,
-            )
-
-            run_post_process_pass(self, _deterministic_sampling_backend)
-            is_deepseek_model = False
-            if parse_connector_type(cfg.model_path) != ConnectorType.INSTANCE:
-                try:
-                    hf_config = self.get_model_config().hf_config
-                    model_arch = hf_config.architectures[0]
-                    is_deepseek_model = model_arch in [
-                        "DeepseekV2ForCausalLM",
-                        "DeepseekV3ForCausalLM",
-                        "DeepseekV32ForCausalLM",
-                        "MistralLarge3ForCausalLM",
-                        "PixtralForConditionalGeneration",
-                        "GlmMoeDsaForCausalLM",
-                        "Glm4MoeLiteForCausalLM",
-                    ]
-                except Exception:
-                    pass
-
-            # Check attention backend
-            run_post_process_pass(self, _deterministic_attention_backend)
-
-            attention_backend = resolved_view(self).attention_backend
-            if is_deepseek_model:
-                if (
-                    attention_backend
-                    not in RADIX_SUPPORTED_DETERMINISTIC_ATTENTION_BACKEND
-                ):
-                    raise ValueError(
-                        f"Currently only {RADIX_SUPPORTED_DETERMINISTIC_ATTENTION_BACKEND} attention backends are supported for deterministic inference with absorbed-MLA models. But you're using {attention_backend}."
-                    )
-                if attention_backend == "fa4" and not is_sm100_or_sm110_supported():
-                    raise ValueError(
-                        "Deterministic inference with absorbed-MLA models on the fa4 "
-                        "attention backend requires SM100/SM110: it runs "
-                        "absorbed MLA, whose qv argument flash_attn.cute only "
-                        "implements on those archs."
-                    )
-
-            if attention_backend not in RADIX_SUPPORTED_DETERMINISTIC_ATTENTION_BACKEND:
-                # Currently, only certain backends support radix cache. Support for other backends is in progress
-                self._declare(
-                    "_handle_deterministic_inference",
-                    disable_radix_cache=True,
-                )
-                logger.warning(
-                    f"Currently radix cache is not compatible with {attention_backend} attention backend for deterministic inference. It will be supported in the future."
-                )
-
-            # Check TP size
-            if cfg.tp_size > 1:
-                if is_hip():
-                    # AMD: use 1-stage all-reduce kernel which is inherently deterministic
-                    # (each GPU reads all data from all GPUs, reduces locally in fixed order)
-                    logger.info(
-                        "AMD/ROCm: Using 1-stage all-reduce kernel (deterministic)"
-                    )
-                else:
-                    # CUDA: use NCCL tree algorithm
-                    os.environ["NCCL_ALGO"] = "allreduce:tree"
-                    # Not declared: set_default_server_args() writes this field
-                    # too, through its `args` parameter, so a declaration here
-                    # would be a second source for one field.
-                    self._declare(
-                        "_handle_deterministic_inference",
-                        disable_custom_all_reduce=True,
-                    )
-                    # should_torch_symm_mem_allreduce() takes the
-                    # symmetric-memory path only below a byte threshold, so
-                    # which reduce runs would follow the token count.
-                    self._declare(
-                        "_handle_deterministic_inference",
-                        enable_torch_symm_mem=False,
-                    )
-                    # Each channel carries a differently shaped tree and the
-                    # channel count is picked from the message size, so a
-                    # token's reduction order would follow the token count.
-                    nchannels = str(envs.SGLANG_DETERMINISTIC_NCCL_NCHANNELS.get())
-                    os.environ["NCCL_MIN_NCHANNELS"] = nchannels
-                    os.environ["NCCL_MAX_NCHANNELS"] = nchannels
-                    logger.warning(
-                        "NCCL_ALGO is set to 'allreduce:tree', the NCCL channel count is pinned, and custom and symmetric-memory all reduce are disabled for deterministic inference when TP size > 1."
-                    )
+        handle_deterministic_inference(self)
 
     def _handle_unified_memory_pool(self):
         from sglang.srt.arg_groups.kv_cache_hook import handle_unified_memory_pool
@@ -7039,103 +5220,9 @@ class ServerArgs:
         handle_page_major_kv_layout(self)
 
     def _handle_dllm_inference(self):
-        cfg = resolving_view(self)
-        if cfg.dllm_algorithm is None:
-            return
-        # On AMD/HIP, disable cuda graph for DLLM (the attention_backend
-        # resolution moved to the pipeline: arg_groups/overrides.py
-        # _dllm_attention_backend, invoked below at its legacy slot).
-        if is_hip():
-            if (
-                cfg.cuda_graph_config.decode.backend != Backend.DISABLED
-                or cfg.cuda_graph_config.prefill.backend != Backend.DISABLED
-            ):
-                logger.warning(
-                    "Cuda graph is disabled for diffusion LLM inference on AMD GPUs"
-                )
-                self._declare(
-                    "_handle_dllm_inference",
-                    cuda_graph_config=with_phase(
-                        cfg.cuda_graph_config, Phase.DECODE, backend=Backend.DISABLED
-                    ),
-                )
-                self._declare(
-                    "_handle_dllm_inference",
-                    cuda_graph_config=with_phase(
-                        cfg.cuda_graph_config, Phase.PREFILL, backend=Backend.DISABLED
-                    ),
-                )
+        from sglang.srt.arg_groups.dllm_hook import handle_dllm_inference
 
-        from sglang.srt.arg_groups.overrides import (
-            _dllm_attention_backend,
-            _dllm_overlap_disable,
-            run_post_process_pass,
-        )
-
-        run_post_process_pass(self, _dllm_attention_backend)
-        run_post_process_pass(self, _dllm_overlap_disable)
-
-        # The page-size alignment + block-size cap for dllm moved to the
-        # resolution pipeline (arg_groups/overrides.py: _dllm_page_size).
-        # Invoked outside the radix gate: the alignment fill keeps its radix
-        # gate inside the pass, the block-size cap applies regardless (it
-        # replaces the unconditional scheduler-init fallback).
-        from sglang.srt.arg_groups.overrides import _dllm_page_size
-
-        run_post_process_pass(self, _dllm_page_size)
-
-        if not cfg.disable_radix_cache:
-            if cfg.enable_hierarchical_cache:
-                logger.warning(
-                    "Hierarchical cache is disabled because of using diffusion LLM inference"
-                )
-                self._declare(
-                    "_handle_dllm_inference",
-                    enable_hierarchical_cache=False,
-                )
-            if cfg.enable_lmcache:
-                logger.warning(
-                    "LMCache is disabled because of using diffusion LLM inference"
-                )
-                self._declare("_handle_dllm_inference", enable_lmcache=False)
-            if cfg.enable_flexkv:
-                logger.warning(
-                    "FlexKV is disabled because of using diffusion LLM inference"
-                )
-                self._declare("_handle_dllm_inference", enable_flexkv=False)
-
-        if cfg.pp_size > 1:
-            logger.warning(
-                "Pipeline parallelism is disabled because of using diffusion LLM inference"
-            )
-            self._declare(
-                "_handle_dllm_inference",
-                pp_size=1,
-            )
-
-        if cfg.enable_lora:
-            logger.warning(
-                "Currently LoRA is not supported by diffusion LLM inference."
-            )
-            self._declare("_handle_dllm_inference", enable_lora=False)
-
-        if cfg.disaggregation_mode != "null":
-            logger.warning(
-                "Currently disaggregation is not supported by diffusion LLM inference."
-            )
-            self._declare(
-                "_handle_dllm_inference",
-                disaggregation_mode="null",
-            )
-
-        if cfg.enable_mixed_chunk:
-            logger.warning(
-                "Mixed chunked prefill is disabled because of using diffusion LLM inference."
-            )
-            self._declare(
-                "_handle_dllm_inference",
-                enable_mixed_chunk=False,
-            )
+        handle_dllm_inference(self)
 
     def _handle_asr_validation(self):
         from sglang.srt.arg_groups.serving_hook import handle_asr_validation
@@ -7148,115 +5235,9 @@ class ServerArgs:
             raise ValueError("--prefill-decode-interval must be non-negative.")
 
     def _handle_other_validations(self):
-        cfg = resolving_view(self)
-        if cfg.default_chat_template_kwargs is not None and not isinstance(
-            cfg.default_chat_template_kwargs, dict
-        ):
-            raise ValueError(
-                "--default-chat-template-kwargs must decode to a JSON object"
-            )
+        from sglang.srt.arg_groups.serving_hook import handle_other_validations
 
-        # Handle optimistic prefill validation
-        if cfg.optimistic_prefill_attempts > 0 and cfg.disaggregation_mode == "prefill":
-            if cfg.pp_size > 1:
-                logger.warning("Optimistic prefill does not support pp_size > 1")
-                self._declare(
-                    "_handle_other_validations",
-                    optimistic_prefill_attempts=0,
-                )
-            elif cfg.enable_hierarchical_cache and (
-                cfg.hicache_storage_backend is not None
-                or cfg.hicache_write_policy != "write_back"
-            ):
-                logger.warning(
-                    "Optimistic prefill only supports L2 hierarchical cache "
-                    "with write-back policy"
-                )
-                self._declare(
-                    "_handle_other_validations",
-                    optimistic_prefill_attempts=0,
-                )
-            elif resolved_view(self).uses_mamba_radix_cache:
-                logger.warning(
-                    "Optimistic prefill does not support models that use "
-                    "mamba radix cache."
-                )
-                self._declare(
-                    "_handle_other_validations",
-                    optimistic_prefill_attempts=0,
-                )
-
-        # Handle model inference tensor dump.
-        if cfg.debug_tensor_dump_output_folder is not None:
-            logger.warning(
-                "Cuda graph and server warmup are disabled because of using tensor dump mode"
-            )
-            self._declare(
-                "_handle_other_validations",
-                cuda_graph_config=with_phase(
-                    cfg.cuda_graph_config, Phase.DECODE, backend=Backend.DISABLED
-                ),
-            )
-            self._declare(
-                "_handle_other_validations",
-                cuda_graph_config=with_phase(
-                    cfg.cuda_graph_config, Phase.PREFILL, backend=Backend.DISABLED
-                ),
-            )
-            self._declare("_handle_other_validations", skip_server_warmup=True)
-
-        if cfg.msprobe_dump_config is not None:
-            logger.warning(
-                "When msProbe is enabled, "
-                "cuda graph is disabled because msProbe only supports dump in eager mode, "
-                "warmup is disabled(skip_server_warmup=True) because there is no need to dump data for this stage."
-            )
-            self._declare(
-                "_handle_other_validations",
-                cuda_graph_config=with_phase(
-                    cfg.cuda_graph_config, Phase.DECODE, backend=Backend.DISABLED
-                ),
-            )
-            self._declare(
-                "_handle_other_validations",
-                cuda_graph_config=with_phase(
-                    cfg.cuda_graph_config, Phase.PREFILL, backend=Backend.DISABLED
-                ),
-            )
-            self._declare("_handle_other_validations", skip_server_warmup=True)
-
-        # Validate limit_mm_per_prompt modalities
-        if cfg.limit_mm_data_per_request:
-            if isinstance(cfg.limit_mm_data_per_request, str):
-                self._declare(
-                    "_handle_other_validations",
-                    limit_mm_data_per_request=json.loads(cfg.limit_mm_data_per_request),
-                )
-
-            if isinstance(cfg.limit_mm_data_per_request, dict):
-                allowed_modalities = {"image", "video", "audio"}
-                for modality in cfg.limit_mm_data_per_request.keys():
-                    if modality not in allowed_modalities:
-                        raise ValueError(
-                            f"Invalid modality '{modality}' in --limit-mm-data-per-request."
-                            f"Allowed modalities are: {list(allowed_modalities)}"
-                        )
-
-        # Validate preferred_sampling_params
-        if cfg.preferred_sampling_params:
-            if isinstance(cfg.preferred_sampling_params, str):
-                self._declare(
-                    "_handle_other_validations",
-                    preferred_sampling_params=json.loads(cfg.preferred_sampling_params),
-                )
-
-            # Validate preferred_sampling_params doesn't use tokenizer-dependent features
-            if cfg.skip_tokenizer_init:
-                from sglang.srt.sampling.sampling_params import SamplingParams
-
-                test_params = SamplingParams(**cfg.preferred_sampling_params)
-                # raises if tokenizer-dependent features used
-                test_params.normalize(None)
+        handle_other_validations(self)
 
     def _handle_crash_dump_env(self):
         from sglang.srt.arg_groups.serving_hook import handle_crash_dump_env
