+++ test/registered/unit/test_supplied_instance_exposure_ratchet.py
@@ -511,12 +511,28 @@ class TestSuppliedInstanceExposure(CustomTestCase):
             "prefill_attention_backend",
             "speculative_draft_attention_backend",
         }
-        deprecated = next(
-            node
-            for node in ast.walk(sa_class)
-            if isinstance(node, ast.FunctionDef)
-            and node.name == "_handle_deprecated_args"
-        )
+
+        # The handler lives in `arg_groups/serving_hook.py` now; the record
+        # keeps a slot that calls it, so the loop is looked up by either name.
+        def _deprecated_alias_handler():
+            for node in ast.walk(sa_class):
+                if (
+                    isinstance(node, ast.FunctionDef)
+                    and node.name == "_handle_deprecated_args"
+                    and any(isinstance(n, ast.For) for n in ast.walk(node))
+                ):
+                    return node
+            for path in sorted((_PACKAGE_ROOT / "arg_groups").glob("*.py")):
+                tree = ast.parse(path.read_text(encoding="utf-8-sig"))
+                for node in tree.body:
+                    if (
+                        isinstance(node, ast.FunctionDef)
+                        and node.name == "handle_deprecated_args"
+                    ):
+                        return node
+            raise AssertionError("the deprecated-alias handler was not found")
+
+        deprecated = _deprecated_alias_handler()
         found_tuples = [
             {elt.value for elt in node.iter.elts if isinstance(elt, ast.Constant)}
             for node in ast.walk(deprecated)
